% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_obs_errs_to_pt_nowcast_mat.R
\name{add_obs_err_to_pt_nowcast_mat}
\alias{add_obs_err_to_pt_nowcast_mat}
\title{Add observation error to a point nowcast}
\usage{
add_obs_err_to_pt_nowcast_mat(point_nowcast_matrix, disp)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of both observations and point estimates
of counts at each reference time and delay.}

\item{disp}{Vector of dispersion parameters indexed starting at delay = 1,}
}
\value{
\code{nowcast_w_obs_error} Matrix containing the same upper left values
as the \code{point_nowcast_matrix}, with the bottom right containing
probabilistic draws assuming a negative binomial
observation model with a mean given by the lower right entries in
\code{point_nowcast_matrix} and dispersion parameters
from \code{disp}
}
\description{
The function ingests an estimate of a point nowcast, in the form of a
reporting rectangle and a vector of dispersion parameters, and adds
observation error to generate an expected observed nowcast
}
\examples{
point_nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 17.8,
    80, 40, 23.2, 15.9,
    70, 35, 20.3, 19.9
  ),
  nrow = 5,
  byrow = TRUE
)

exp_obs_nowcast <- add_obs_err_to_pt_nowcast_mat(
  point_nowcast_matrix = point_nowcast_matrix,
  disp = c(8, 1.4, 4)
)
print(exp_obs_nowcast)
}
