% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_obs_errors_to_nowcast.R
\name{add_obs_errors_to_nowcast}
\alias{add_obs_errors_to_nowcast}
\title{Add observation errors to nowcasts}
\usage{
add_obs_errors_to_nowcast(comp_rep_square, disp, n_draws = 1000)
}
\arguments{
\item{comp_rep_square}{Matrix containing both observed and point estimated
values of the reporting triangle, i.e. the complete
reporting square}

\item{disp}{Vector of dispersion parameters of a negative binomial, for
delays from 1 through the maximum delay}

\item{n_draws}{Integer indicating the number of expected observed reporting
squares to generate}
}
\value{
\code{list_of_exp_obs_nowcasts} List of \code{n_draws} matrices of expected
obsrved reporting sqaures
}
\description{
This function ingests a point nowcast in the form of a completed reporting
square and a vector of dispersion parameters. It generates a list of
\code{n_draws} expected observed completed reporting squares.
}
\examples{
point_nowcast <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 17.8,
    80, 40, 23.2, 15.9,
    70, 35, 20.3, 19.9
  ),
  nrow = 5,
  byrow = TRUE
)

list_of_exp_obs_nowcast <- add_obs_errors_to_nowcast(
  comp_rep_square = point_nowcast,
  disp = c(8, 1.4, 4),
  n_draws = 10
)
print(list_of_exp_obs_nowcast[[1]])
}
