% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_df_by_ref_time.R
\name{aggregate_df_by_ref_time}
\alias{aggregate_df_by_ref_time}
\title{Summarise a dataframe containing draws of reporting squares}
\usage{
aggregate_df_by_ref_time(nowcast_draws_df)
}
\arguments{
\item{nowcast_draws_df}{Dataframe containing draws of a reporting square.
Must contain the following columns: \code{time}, \code{delay}, \code{draw}, \code{count}.}
}
\value{
\code{summary_df} Dataframe containing the nowcasted expected observed
total observations at each reference time.
}
\description{
This function ingests a long tidy dataframe containing many draws of
reporting squares across each reference time, by summing over all the
counts reported at each delay \code{d}.
}
\examples{
nowcast_draws_df <- data.frame(
  time = c(1, 1, 1, 1, 2, 2, 2, 2),
  delay = c(1, 2, 1, 2, 1, 2, 1, 2),
  draw = c(1, 1, 2, 2, 1, 1, 2, 2),
  count = c(3, 6, 4, 7, 1, 2, 2, 3)
)
summary_df <- aggregate_df_by_ref_time(nowcast_draws_df)
print(summary_df)
}
