% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_delay.R
\name{apply_delay}
\alias{apply_delay}
\title{Apply the delay to generate a point nowcast}
\usage{
apply_delay(triangle_to_nowcast, delay_pmf)
}
\arguments{
\item{triangle_to_nowcast}{matrix of the incomplete reporting triangle to be
nowcasted, with rows representing the time points of reference and columns
representing the delays}

\item{delay_pmf}{vector of delays assumed to be indexed starting at the
first delay column in \code{triangle_to_nowcast}}
}
\value{
expectation, a matrix of the same number of rows and columns as the
\code{triangle_to_nowcast} but with the missing values filled in as point
estimates
}
\description{
This function takes as an input the reporting triangle that we want to
complete with a nowcast and a delay pmf and generates a point estimate of a
completed reporting square (or rectangle)
}
\examples{
library(epinowcast)
nat_germany_hosp <-
  germany_covid19_hosp[location == "DE"][age_group == "00+"]
nat_germany_hosp <- enw_filter_report_dates(
  nat_germany_hosp,
  latest_date = "2021-10-01"
)
pobs <- enw_preprocess_data(nat_germany_hosp, max_delay = 21)
triangle_raw <- pobs$reporting_triangle[[1]] |>
  dplyr::select(-`.group`, -reference_date) |>
  as.matrix() |>
  unname()
delay_df <- get_delay_estimate(
  triangle = triangle_raw,
  max_delay = 20,
  n_history = 30
)
reporting_square <- apply_delay(
  triangle_to_nowcast = triangle_raw,
  delay_pmf = delay_df$pmf
)
}
