% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as_data_frame}
\alias{as_data_frame}
\alias{as_data_frame.matrix}
\title{Create a long tidy dataframe from a reporting triangle}
\usage{
as_data_frame(data, reference_dates = NULL, delays = NULL, ...)

\method{as_data_frame}{matrix}(data, reference_dates, delays, ...)
}
\arguments{
\item{data}{Reporting triangle to be converted to a long tidy dataframe
with counts by reference and report date. Can either be a matrix or a
dataframe but should be in wide format where rows are reference times and
columns are delays plus any additional metadata.}

\item{reference_dates}{Vector of dates or character strings indicating the
dates that each of the reference times correspond to, ordered from
row 1 to the last row of the reporting matrix. Must be of the same length
as the number of rows of the reporting matrix.}

\item{delays}{Vector of integers indicating the delays that each column of
the reporting matrix corresponds to.}

\item{...}{Additional arguments}
}
\value{
Data.frame containing the counts by reference date and report
date
}
\description{
Create a long tidy dataframe from a reporting triangle
}
\details{
The input needs to be a matrix or data.frame with numeric or NA entries
}
\examples{
rep_tri <- matrix(
  c(
    10, 2, 1,
    12, 3, NA,
    13, NA, NA
  ),
  byrow = TRUE,
  nrow = 3
)
reference_dates <- c("2025-01-01", "2025-01-02", "2025-01-03")
delays <- c(0, 1, 2)
as_data_frame(
  rep_tri,
  reference_dates,
  delays
)
}
