% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{as_reporting_triangle.data.frame}
\alias{as_reporting_triangle.data.frame}
\title{Create a \code{reporting_triangle} object from a data.frame}
\usage{
\method{as_reporting_triangle}{data.frame}(
  data,
  max_delay,
  strata = NULL,
  delays_unit = "days",
  reference_date = "reference_date",
  report_date = "report_date",
  count = "count",
  ...
)
}
\arguments{
\item{data}{Data.frame in a long tidy format with counts by reference date
and report date. Must contain the following columns:
.    - Column of type \code{date} or character with the dates of
the primary event occurrence (reference date).
\itemize{
\item Column of type \code{date} or character with the dates of
report of the primary event (report_date).
\item Column of numeric or integer indicating the new confirmed counts
pertaining to that reference and report date (count).
Additional columns can be included but will not be used. The input
dataframe for this function must contain only a single strata, there can
be no repeated reference dates and report dates.
}}

\item{reference_date}{Character string indicating the name of the
column which represents the reference date, or the date of the primary
event occurrence.}

\item{report_date}{Character string indicating the name of the
column which represents the date the primary event was reported.}

\item{count}{Character string indicating the name of the column
containing the number of incident cases on each reference and report date.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{reporting_triangle} object.
}
\description{
This method takes a data.frame containing case counts indexed by reference
date and report date and creates a \code{reporting_triangle} object. See other
methods for other data input options.
}
\examples{
data_as_of_df <- syn_nssp_df[syn_nssp_df$report_date <= "2026-04-01", ]
as_reporting_triangle(
  data = data_as_of_df,
  max_delay = 25
)
}
\seealso{
\code{\link{as_reporting_triangle}}
\code{\link{as_reporting_triangle.matrix}}
}
