% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{as_reporting_triangle.matrix}
\alias{as_reporting_triangle.matrix}
\title{Create a \code{reporting_triangle} from a matrix}
\usage{
\method{as_reporting_triangle}{matrix}(data, reference_dates, delays_unit = "days", ...)
}
\arguments{
\item{data}{Matrix of a reporting triangle where rows are reference times,
columns are delays, and entries are the incident counts. The number of
columns determines the maximum delay (ncol - 1).}

\item{reference_dates}{Vector of character strings indicating the reference
dates corresponding to each row of the reporting triangle matrix (\code{data}).}

\item{delays_unit}{Character string specifying the temporal granularity of
the delays. Options are \code{"days"}, \code{"weeks"}, \code{"months"}, \code{"years"}.
For the matrix method, this is simply passed as an item in the
\code{reporting_triangle} object which will later be used to create a nowcast
data.frame. For the data.frame method, this is used to compute the delay
in terms of the specified unit, and to expand the combinations of
reference dates and delays to the complete set of combinations in
the reporting triangle.  Default is \code{"days"}.}

\item{...}{Additional arguments not used.}
}
\value{
A \code{\link{reporting_triangle}} object
}
\description{
This method takes a matrix in the format of a reporting triangle, with rows
as reference dates and columns as delays and elements as incident case
counts and creates a \link{reporting_triangle} object. The maximum delay is
automatically inferred from the number of columns in the matrix. See
\code{\link[=as_reporting_triangle.data.frame]{as_reporting_triangle.data.frame()}} for other data input options.
}
\examples{
rep_tri_mat <- matrix(
  c(
    1, 3, 5, 7, 9,
    4, 7, 8, 0, NA,
    9, 10, 0, NA, NA,
    3, 0, NA, NA, NA,
    6, NA, NA, NA, NA
  ),
  nrow = 5,
  byrow = TRUE
)

reference_dates <- seq(
  from = as.Date("2025-01-01"),
  to = as.Date("2025-01-05"),
  by = "day"
)
rep_tri <- as_reporting_triangle(
  data = rep_tri_mat,
  reference_dates = reference_dates
)
rep_tri

# Access the computed max_delay
get_max_delay(rep_tri)
}
\seealso{
Reporting triangle construction and validation
\code{\link{as_ChainLadder_triangle}()},
\code{\link{as_reporting_triangle}()},
\code{\link{as_reporting_triangle.data.frame}()},
\code{\link{as_reporting_triangle.triangle}()},
\code{\link{assert_reporting_triangle}()},
\code{\link{detect_structure}()},
\code{\link{new_reporting_triangle}()},
\code{\link{reporting_triangle-class}}
}
\concept{reporting_triangle}
