% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle_df.R
\name{as_reporting_triangle_df.data.frame}
\alias{as_reporting_triangle_df.data.frame}
\title{Create a \code{reporting_triangle_df} object from a data.frame}
\usage{
\method{as_reporting_triangle_df}{data.frame}(
  data,
  delays_unit = "days",
  reference_date = "reference_date",
  report_date = "report_date",
  count = "count",
  by = NULL,
  ...
)
}
\arguments{
\item{data}{Data.frame in a long tidy format with counts by reference date
and report date. Must contain the following columns:
\itemize{
\item Column of type \code{Date} or character with the dates of
the primary event occurrence (reference date).
\item Column of type \code{Date} or character with the dates of
report of the primary event (report_date).
\item Column of numeric or integer indicating the new confirmed counts
pertaining to that reference and report date (count).
Additional columns can be included and will be preserved if specified in
the \code{by} parameter as strata columns.
}}

\item{delays_unit}{Character string specifying the temporal granularity of
the delays. Options are \code{"days"}, \code{"weeks"}, \code{"months"}, \code{"years"}.
Default is \code{"days"}.}

\item{reference_date}{Character string indicating the name of the
column which represents the reference date, or the date of the primary
event occurrence. Default is "reference_date".}

\item{report_date}{Character string indicating the name of the
column which represents the date the primary event was reported.
Default is "report_date".}

\item{count}{Character string indicating the name of the column
containing the number of incident cases on each reference and report date.
Default is "count".}

\item{by}{Character vector of strata column names. These columns define
unique strata for nowcasting. NULL (default) indicates no strata.
Stored internally as the "strata" attribute.}

\item{...}{Additional arguments not used.}
}
\value{
A \link{reporting_triangle_df} object
}
\description{
This method takes a data.frame containing case counts indexed by reference
date and report date and creates a \link{reporting_triangle_df} object.
}
\examples{
# Single stratum (no strata)
data_as_of_df <- syn_nssp_df[syn_nssp_df$report_date <= "2026-04-01", ]
rt_df <- as_reporting_triangle_df(data = data_as_of_df)

# Multiple strata
rt_df_strata <- as_reporting_triangle_df(
  germany_covid19_hosp,
  by = c("age_group", "location")
)
}
\seealso{
Other reporting_triangle_df:
\code{\link{[.reporting_triangle_df}()},
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{assert_reporting_triangle_df}()},
\code{\link{get_strata}()},
\code{\link{head.reporting_triangle_df}()},
\code{\link{is_reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{print.reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{summary.reporting_triangle_df}()},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{validate_reporting_triangle_df}()}
}
\concept{reporting_triangle_df}
