% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle_df.R
\name{as_reporting_triangle_df}
\alias{as_reporting_triangle_df}
\title{Create a \code{reporting_triangle_df} object}
\usage{
as_reporting_triangle_df(data, delays_unit = "days", ...)
}
\arguments{
\item{data}{Data to be converted to reporting_triangle_df format.}

\item{delays_unit}{Character string specifying the temporal granularity of
the delays. Options are \code{"days"}, \code{"weeks"}, \code{"months"}, \code{"years"}.
Default is \code{"days"}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \link{reporting_triangle_df} object
}
\description{
Create a \code{reporting_triangle_df} object
}
\seealso{
Other reporting_triangle_df: 
\code{\link{[.reporting_triangle_df}()},
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.data.frame}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{assert_reporting_triangle_df}()},
\code{\link{get_strata}()},
\code{\link{head.reporting_triangle_df}()},
\code{\link{is_reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{print.reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{summary.reporting_triangle_df}()},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{truncate_to_delay}()},
\code{\link{truncate_to_quantile}()},
\code{\link{validate_reporting_triangle_df}()}
}
\concept{reporting_triangle_df}
