% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_delays_unit}
\alias{assert_delays_unit}
\title{Assert delays_unit is valid}
\usage{
assert_delays_unit(delays_unit)
}
\arguments{
\item{delays_unit}{Character string specifying the temporal granularity}
}
\value{
NULL, invisibly. Stops execution with error if validation fails.
}
\description{
Assert delays_unit is valid
}
\keyword{internal}
