% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{assert_nowcast_df}
\alias{assert_nowcast_df}
\title{Assert validity of \code{nowcast_df} objects}
\usage{
assert_nowcast_df(data)
}
\arguments{
\item{data}{A \code{\link{nowcast_df}} object to check for validity.}
}
\description{
Assert validity of \code{nowcast_df} objects
}
