% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{assert_reporting_triangle}
\alias{assert_reporting_triangle}
\title{Assert validity of \code{reporting_triangle} objects}
\usage{
assert_reporting_triangle(data, ...)
}
\arguments{
\item{data}{An object to check for validity.}

\item{...}{Additional arguments}
}
\description{
Assert validity of \code{reporting_triangle} objects
}
