% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{assert_reporting_triangle.reporting_triangle}
\alias{assert_reporting_triangle.reporting_triangle}
\title{S3 method for reporting triangle assertion}
\usage{
\method{assert_reporting_triangle}{reporting_triangle}(data, ...)
}
\arguments{
\item{data}{An object to check for validity.}

\item{...}{Additional arguments}
}
\description{
S3 method for reporting triangle assertion
}
