% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle_df-class.R
\name{assert_reporting_triangle_df}
\alias{assert_reporting_triangle_df}
\title{Assert validity of \code{reporting_triangle_df} objects}
\usage{
assert_reporting_triangle_df(data, validate = TRUE)
}
\arguments{
\item{data}{A \link{reporting_triangle_df} object to check for validity.}

\item{validate}{Logical. If TRUE (default), validates the object. Set to
FALSE only when called from functions that already validated.}
}
\description{
Assert validity of \code{reporting_triangle_df} objects
}
\seealso{
Other reporting_triangle_df:
\code{\link{[.reporting_triangle_df}()},
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.data.frame}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{get_strata}()},
\code{\link{head.reporting_triangle_df}()},
\code{\link{is_reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{print.reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{summary.reporting_triangle_df}()},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{validate_reporting_triangle_df}()}
}
\concept{reporting_triangle_df}
