% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_uncertainty_model.R
\name{assert_uncertainty_model.default}
\alias{assert_uncertainty_model.default}
\title{Default Validation for Uncertainty Models}
\usage{
\method{assert_uncertainty_model}{default}(model, ...)
}
\arguments{
\item{model}{Object to validate.}

\item{...}{Additional arguments passed to methods.}
}
\value{
NULL (invisibly) if valid, otherwise throws an error.
}
\description{
Default method that throws an error if the object is not an
uncertainty_model.
}
\seealso{
Other uncertainty_model: 
\code{\link{assert_uncertainty_model}()},
\code{\link{assert_uncertainty_model.uncertainty_model}()},
\code{\link{assert_uncertainty_model.uncertainty_nb}()},
\code{\link{assert_uncertainty_model.uncertainty_poisson}()},
\code{\link{print.uncertainty_model}()},
\code{\link{uncertainty_model}()}
}
\concept{uncertainty_model}
