% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_uncertainty_model.R
\name{assert_uncertainty_model.uncertainty_poisson}
\alias{assert_uncertainty_model.uncertainty_poisson}
\title{Validate Poisson Uncertainty Model}
\usage{
\method{assert_uncertainty_model}{uncertainty_poisson}(model, ...)
}
\arguments{
\item{model}{Object to validate.}

\item{...}{Additional arguments passed to methods.}
}
\value{
NULL (invisibly) if valid, otherwise throws an error.
}
\description{
Validates that the model is correctly configured as a Poisson
uncertainty model.
}
\seealso{
Other uncertainty_model: 
\code{\link{assert_uncertainty_model}()},
\code{\link{assert_uncertainty_model.default}()},
\code{\link{assert_uncertainty_model.uncertainty_model}()},
\code{\link{assert_uncertainty_model.uncertainty_nb}()},
\code{\link{print.uncertainty_model}()},
\code{\link{uncertainty_model}()}
}
\concept{uncertainty_model}
