% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_strategy.R
\name{assert_uncertainty_strategy}
\alias{assert_uncertainty_strategy}
\title{Validate Uncertainty Strategy}
\usage{
assert_uncertainty_strategy(strategy)
}
\arguments{
\item{strategy}{Object to validate.}
}
\value{
NULL (invisibly) if valid, otherwise throws an error.
}
\description{
Internal function to validate that an object is a valid
uncertainty_strategy.
}
\keyword{internal}
