% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{baselinenowcast.data.frame}
\alias{baselinenowcast.data.frame}
\title{Create a dataframe of nowcast results from a dataframe of cases
indexed by reference date and report date}
\usage{
\method{baselinenowcast}{data.frame}(
  data,
  scale_factor = 3,
  prop_delay = 0.5,
  output_type = c("samples", "point"),
  draws = 1000,
  uncertainty_model = fit_by_horizon,
  uncertainty_sampler = sample_nb,
  max_delay,
  delays_unit = "days",
  reference_date = "reference_date",
  report_date = "report_date",
  count = "count",
  nowcast_unit = NULL,
  strata_sharing = NULL,
  delay_pmf = NULL,
  uncertainty_params = NULL,
  ...
)
}
\arguments{
\item{data}{Data.frame in a long tidy format with counts by reference date
and report date for one or more strata. Must contain the following
columns:
.    - Column of type \code{date} or character with the dates of
the primary event occurrence (reference date).
\itemize{
\item Column of type \code{date} or character with the dates of
report of the primary event (report_date).
\item Column of numeric or integer indicating the new confirmed counts
pertaining to that reference and report date (count).
Additional columns can be included, and the user can specify which columns
set the unit of a single nowcast  ( i.e. the combination of columns that
uniquely define a single nowcast with the \code{nowcast_unit} argument.).
}}

\item{scale_factor}{Numeric value indicating the multiplicative factor on
the maximum delay to be used for estimation of delay and uncertainty.
Default is \code{3}.}

\item{prop_delay}{Numeric value <1 indicating what proportion of all
reference times in the reporting triangle to be used for delay
estimation. Default is \code{0.5}.}

\item{output_type}{Character string indicating whether the output should be
samples (\code{"samples"})from the estimate with full uncertainty or whether to
return the point estimate (\code{"point"}). Default is \code{"samples"}.}

\item{draws}{Integer indicating the number of probabilistic draws to include
if \code{output_type} is \code{"samples"}. Default is 1000.}

\item{uncertainty_model}{Function that ingests a matrix of observations and a
matrix of predictions and returns a vector that can be used to
apply uncertainty using the same error model. Default is
\code{fit_by_horizon} with arguments of \code{obs} matrix of observations and
\code{pred} the matrix of predictions that fits each column (horizon)
to a negative binomial observation model by default. The user can
specify a different fitting model by replacing the
\code{fit_model} argument in \code{fit_by_horizon}.}

\item{uncertainty_sampler}{Function that ingests a vector or matrix of
predictions and a vector of uncertainty parameters and generates draws
from the observation model. Default is \code{sample_nb} which expects
arguments \code{pred} for the vector of predictions and uncertainty parameters
for the corresponding vector of uncertainty parameters, and draws from a
negative binomial for each element of the vector.}

\item{max_delay}{Integer indicating the maximum delay.}

\item{delays_unit}{Character string specifying the temporal granularity of
the delays. Options are \code{"days"}, \code{"weeks"}, \code{"months"}, \code{"years"}.
For the matrix method, this is simply passed as an item in the
\code{reporting_triangle} object which will later be used to create a nowcast
data.frame. For the data.frame method, this is used to compute the delay
in terms of the specified unit, and to expand the combinations of
reference dates and delays to the complete set of combinations in
the reporting triangle.  Default is \code{"days"}.}

\item{reference_date}{Character string indicating the name of the
column which represents the reference date, or the date of the primary
event occurrence.}

\item{report_date}{Character string indicating the name of the
column which represents the date the primary event was reported.}

\item{count}{Character string indicating the name of the column
containing the number of incident cases on each reference and report date.}

\item{nowcast_unit}{Vector of character strings indicating the names of the
columns in \code{data} (after any renaming of columns) that denote the unit of
a single nowcast.Within a nowcast unit, there can be no repeated unique
combinations of reference dates and report dates. Default is \code{NULL} which
assumes that all columns that are not required columns form the unit of
a single forecast. This may lead to unexpected behaviour, so setting the
nowcast unit explicitly can help make the code easier to debug and easier
to read. If specified, all columns that are not part of the forecast unit
(or required columns) will be removed.}

\item{strata_sharing}{Vector of character strings indicating the estimand
for which estimates that are "borrowed" from across all strata should
be used. Options are \code{"delay"} and/or \code{"uncertainty"}. NULL indicates that
delay and uncertainty estimates should be computed for each \code{nowcast_unit}
independently.}

\item{delay_pmf}{Vector of delays assumed to be indexed starting at the
first delay. Default is NULL,
which will estimate the delay from the reporting triangle in each
nowcast unit in \code{data}. See \code{\link{estimate_delay}} for more details.}

\item{uncertainty_params}{Vector of uncertainty parameters ordered from
horizon 1 to the maximum horizon. Default is \code{NULL}, which will
estimate the uncertainty parameters from the reporting triangle in each
nowcast unit in \code{data}. See \code{\link{estimate_uncertainty}} for more
details.}

\item{...}{Additional arguments passed to
\code{\link{estimate_uncertainty}}
and \code{\link{sample_nowcast}}.}
}
\value{
Data.frame of class \code{\link{baselinenowcast_df}}
}
\description{
This function ingests a data.frame with the number of incident
cases indexed by reference date and report date for potentially multiple
strata (e.g. age groups or locations) and returns a data.frame containing
nowcasts by reference dates for each of the specified strata.
For each strata, this function will by default estimate uncertainty using
past retrospective nowcast errors and generate probabilistic nowcasts,
which are samples from the predictive distribution of the estimated final
case count at each reference date. See documentation for the arguments of
this function which can be used to set the model specifications (things
like number of reference times for delay and uncertainty estimation,
the observation model, etc.). The function expects that each strata in
the dataframe has the same maximum delay and the same number of reference
dates.
}
\examples{
nowcasts_df <- baselinenowcast(germany_covid19_hosp,
  max_delay = 40,
  nowcast_unit = c("age_group", "location")
)
nowcasts_df
}
\seealso{
Other baselinenowcast_df:
\code{\link{baselinenowcast}()},
\code{\link{baselinenowcast.reporting_triangle}()},
\code{\link{baselinenowcast_df-class}}
}
\concept{baselinenowcast_df}
