% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{baselinenowcast.default}
\alias{baselinenowcast.default}
\title{Default method for baselinenowcast with helpful error messages}
\usage{
\method{baselinenowcast}{default}(data, ...)
}
\arguments{
\item{data}{Data object of unsupported type.}

\item{...}{Additional arguments (not used).}
}
\description{
Provides helpful error messages when an unsupported data type
is passed to \code{\link[=baselinenowcast]{baselinenowcast()}}.
}
\seealso{
Main nowcasting interface functions
\code{\link{assert_baselinenowcast_df}()},
\code{\link{baselinenowcast}()},
\code{\link{baselinenowcast.data.frame}()},
\code{\link{baselinenowcast.reporting_triangle}()},
\code{\link{baselinenowcast.reporting_triangle_df}()},
\code{\link{baselinenowcast_df-class}},
\code{\link{new_baselinenowcast_df}()}
}
\concept{baselinenowcast_df}
