% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{baselinenowcast}
\alias{baselinenowcast}
\alias{baselinenowcast.reporting_triangle}
\title{Create a reporting triangle}
\usage{
baselinenowcast(
  data,
  delay_pmf = NULL,
  uncertainty_params = NULL,
  scale_factor = 3,
  prop_delay = 0.5,
  include_draws = TRUE,
  draws = 1000,
  ...
)

\method{baselinenowcast}{reporting_triangle}(
  data,
  delay_pmf = NULL,
  uncertainty_params = NULL,
  scale_factor = 3,
  prop_delay = 0.5,
  include_draws = TRUE,
  draws = 1000,
  uncertainty_model = fit_by_horizon,
  uncertainty_sampler = sample_nb,
  ...
)
}
\arguments{
\item{data}{Either a \code{reporting_triangle} object or a \code{reporting_triangles}
object.}

\item{delay_pmf}{Vector of delays assumed to be indexed starting at the
first delay column in \code{rep_mat_to_nowcast}}

\item{uncertainty_params}{Vector of uncertainty parameters ordered from
horizon 1 to the maximum horizon. Note that these will be reversed
internally to match the ordering of the \code{point_nowcast_matrix} (where
a horizon of 1 is the last entry).}

\item{scale_factor}{Numeric value indicating the multiplicative factor on
the maximum delay to be used for estimation of delay and uncertainty.
Default is \code{3}.}

\item{prop_delay}{Numeric value <1 indicating what proportion of all
reference times in the reporting triangle to be used for delay
estimation. Default is \code{0.5}.}

\item{include_draws}{Boolean indicating whether or not to include uncertainty
in the output, default is TRUE.}

\item{draws}{Integer indicating the number of probabilistic draws to include
if \code{include_draws} is TRUE. Default is 1000.}

\item{...}{Additional arguments passed to methods. For the
\code{reporting_triangles} method: \code{strata}, \code{strata_sharing}.}

\item{uncertainty_model}{Function that ingests a matrix of observations and a
matrix of predictions and returns a vector that can be used to
apply uncertainty using the same error model. Default is
\code{fit_by_horizon} with arguments of \code{obs} matrix of observations and
\code{pred} the matrix of predictions that fits each column (horizon)
to a negative binomial observation model by default. The user can
specify a different fitting model by replacing the
\code{fit_model} argument in \code{fit_by_horizon}.}

\item{uncertainty_sampler}{Function that ingests a vector or matrix of
predictions and a vector of uncertainty parameters and generates draws
from the observation model. Default is \code{sample_nb} which expects
arguments \code{pred} for the vector of predictions and uncertainty parameters
for the corresponding vector of uncertainty parameters, and draws from a
negative binomial for each element of the vector.}
}
\value{
\code{reporting_triangle} class object which is a list containing:
\itemize{
\item A matrix with which rows are reference times and columns are delays and
entries are incident cases at each reference time and delay.
\item An integer indicating the maximum delay used to create the reporting
triangle
\item A vector of the same length as the rows of the matrix indicating the
dates corresponding to the reference times in the rows of the reporting
triangle.
\item A character string indicating the strata.
\item A vector indicating the "structure" of the reporting triangle.
\item A character string indicating the unit of the delays.
}
}
\description{
Create a reporting triangle

Creating a dataframe of nowcast results from a single reporting
triangle
}
\examples{
data_as_of_df <- syn_nssp_df[syn_nssp_df$report_date <= "2026-04-01", ]
rep_tri <- as_reporting_triangle(
  data = data_as_of_df,
  max_delay = 25
)
nowcast_obj <- baselinenowcast(rep_tri)
}
