% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{baselinenowcast.reporting_triangle}
\alias{baselinenowcast.reporting_triangle}
\title{Creating a dataframe of nowcast results from a single reporting
triangle}
\usage{
\method{baselinenowcast}{reporting_triangle}(
  data,
  delay_pmf = NULL,
  uncertainty_params = NULL,
  scale_factor = 3,
  prop_delay = 0.5,
  include_draws = TRUE,
  draws = 1000,
  uncertainty_model = fit_by_horizon,
  uncertainty_sampler = sample_nb,
  ...
)
}
\arguments{
\item{data}{\code{reporting_triangle} class object to be nowcasted.}

\item{delay_pmf}{Vector of delays assumed to be indexed starting at the
first delay column in \code{rep_mat_to_nowcast}}

\item{uncertainty_params}{Vector of uncertainty parameters ordered from
horizon 1 to the maximum horizon. Note that these will be reversed
internally to match the ordering of the \code{point_nowcast_matrix} (where
a horizon of 1 is the last entry).}

\item{scale_factor}{Numeric value indicating the multiplicative factor on
the maximum delay to be used for estimation of delay and uncertainty.
Default is \code{3}.}

\item{prop_delay}{Numeric value <1 indicating what proportion of all
reference times in the reporting triangle to be used for delay
estimation. Default is \code{0.5}.}

\item{include_draws}{Boolean indicating whether or not to include uncertainty
in the output, default is TRUE.}

\item{draws}{Integer indicating the number of probabilistic draws to include
if \code{include_draws} is TRUE. Default is 1000.}

\item{...}{Additional arguments passed to
\verb{\code{\link{estimate_uncertainty}}}.}
}
\value{
A data.frame of class \code{nowcast_df}
}
\description{
Creating a dataframe of nowcast results from a single reporting
triangle
}
\examples{
data_as_of_df <- syn_nssp_df[syn_nssp_df$report_date <= "2026-04-01", ]
rep_tri <- as_reporting_triangle(
  data = data_as_of_df,
  max_delay = 25
)
nowcast_df<- baselinenowcast(rep_tri)
}
\seealso{
Other nowcast_df:
\code{\link{baselinenowcast}()}
}
\concept{nowcast_df}
