% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_predictions.R
\name{combine_obs_with_pred}
\alias{combine_obs_with_pred}
\title{Combine observed data with a single prediction draw}
\usage{
combine_obs_with_pred(
  predicted_counts,
  reporting_triangle,
  uncertainty = uncertainty_opts(),
  ref_time_aggregator = NULL,
  delay_aggregator = NULL
)
}
\arguments{
\item{predicted_counts}{Vector of predicted counts at each reference time.
Note that if using a reference time or delay aggregator function, this
is assumed to have already been aggregated.}

\item{reporting_triangle}{Reporting triangle matrix.}

\item{uncertainty}{An object of class \code{uncertainty_opts} created by
\code{\link[=uncertainty_opts]{uncertainty_opts()}}. Specifies the aggregation functions. Default uses
standard observed aggregation.}

\item{ref_time_aggregator}{(Deprecated) Function that operates along the
rows (reference times). Use \code{uncertainty} parameter instead.}

\item{delay_aggregator}{(Deprecated) Function that operates along the
columns (delays). Use \code{uncertainty} parameter instead.}
}
\value{
A vector of predicted counts at each reference time
}
\description{
Internally it sums observed counts from the reporting triangle by reference
time and adds them to the predicted counts to form a single draw of the
nowcast for the final counts by reference time.
}
\examples{
pred_counts <- c(10, 20, 30, 40)
reporting_matrix <- matrix(
  c(
    7, 9, 4, 3,
    1, 2, 3, 4,
    5, 6, 7, 8,
    9, 10, 11, 12
  ),
  nrow = 4,
  byrow = TRUE
)
reporting_triangle <- construct_triangle(reporting_matrix)
combine_obs_with_pred(pred_counts, reporting_triangle)
# Another example with rolling sum
if (requireNamespace("zoo", quietly = TRUE)) {
  combine_obs_with_pred(pred_counts,
    reporting_triangle,
    uncertainty = uncertainty_opts(
      aggregation = aggregation_opts(
        ref_time = function(x) zoo::rollsum(x, k = 2, align = "right")
      )
    )
  )
}
}
\seealso{
Probabilistic nowcast generation functions
\code{\link{sample_nb}()},
\code{\link{sample_nowcast}()},
\code{\link{sample_nowcasts}()},
\code{\link{sample_prediction}()},
\code{\link{sample_predictions}()}
}
\concept{generate_probabilistic_nowcasts}
