% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_triangles.R
\name{construct_triangles}
\alias{construct_triangles}
\title{Generate retrospective reporting triangles}
\usage{
construct_triangles(
  truncated_reporting_triangles,
  structure = 1,
  validate = TRUE
)
}
\arguments{
\item{truncated_reporting_triangles}{List of \code{n} truncated reporting triangle
matrices with as many rows as available given the truncation.}

\item{structure}{Integer or vector specifying the reporting structure.
If integer, divides columns evenly by that integer (with last possibly
truncated).  If vector, the sum must not be greater than or equal to the
number of columns. Default is 1 (standard triangular structure).}

\item{validate}{Logical. If TRUE (default), validates the object. Set to
FALSE only when called from functions that already validated.}
}
\value{
\code{reporting_triangles} List of retrospective reporting triangles,
generated by removing the bottom right observations from the truncated
reporting triangle matrices.
}
\description{
This function ingests a list of truncated reporting triangles and iteratively
generates the reporting triangle that would have been available as of the
maximum reference time. It operates on each element in the list in order
(from most recent retrospective nowcast time to oldest retrospective
nowcast time).
}
\examples{
# Generate retrospective triangles from truncated triangles
trunc_rts <- truncate_to_rows(example_reporting_triangle, n = 2)
retro_rts <- construct_triangles(trunc_rts)

# With custom structure
retro_rts_custom <- construct_triangles(
  trunc_rts,
  structure = 2
)
retro_rts_custom
}
\seealso{
Retrospective data generation functions
\code{\link{construct_triangle}()},
\code{\link{truncate_to_row}()},
\code{\link{truncate_to_rows}()}
}
\concept{generate_retrospective_data}
