% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{delay_aggregator_default}
\alias{delay_aggregator_default}
\title{Delay aggregator default function}
\usage{
delay_aggregator_default(fun = rowSums, ...)
}
\arguments{
\item{fun}{Function to aggregate across columns (delays) of the reporting
triangle. Default is \code{rowSums}.}

\item{...}{Additional arguments passed to the function}
}
\value{
Object of class aggregator containing functions and arguments
}
\description{
Delay aggregator default function
}
\examples{
delay_aggregator()
}
