% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_structure.R
\name{detect_structure}
\alias{detect_structure}
\title{Detect the structure of a reporting triangle}
\usage{
detect_structure(reporting_triangle)
}
\arguments{
\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}
}
\value{
Integer or vector specifying the reporting structure.
If integer, divides columns evenly by that integer (with last possibly
truncated).  If vector, the sum must not be greater than or equal to the
number of columns. Default is 1 (standard triangular structure). If
there are no NAs, will return 0.
}
\description{
Detect the structure of a reporting triangle
}
\examples{
ragged_triangle <- matrix(
  c(
    1, 3, 5, 7, 9, 7,
    4, 5, 9, 4, NA, NA,
    1, 6, NA, NA, NA, NA,
    3, NA, NA, NA, NA, NA
  ),
  nrow = 4,
  byrow = TRUE
)
detected_structure <- detect_structure(ragged_triangle)
detected_structure
}
