% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-getters.R
\name{.add_days}
\alias{.add_days}
\title{Internal: Add days to dates}
\usage{
.add_days(dates, delays)
}
\arguments{
\item{dates}{Date vector}

\item{delays}{Numeric vector of delays}
}
\value{
Date vector
}
\description{
Internal: Add days to dates
}
\keyword{internal}
