% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-getters.R
\name{.add_months}
\alias{.add_months}
\title{Internal: Add months to dates}
\usage{
.add_months(dates, delays)
}
\arguments{
\item{dates}{Date vector}

\item{delays}{Numeric vector of delays}
}
\value{
Date vector
}
\description{
Internal: Add months to dates
}
\keyword{internal}
