% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-getters.R
\name{.add_years}
\alias{.add_years}
\title{Internal: Add years to dates}
\usage{
.add_years(dates, delays)
}
\arguments{
\item{dates}{Date vector}

\item{delays}{Numeric vector of delays}
}
\value{
Date vector
}
\description{
Internal: Add years to dates
}
\keyword{internal}
