% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{.apply_aggregator}
\alias{.apply_aggregator}
\title{Internal helper function to apply aggregator to data}
\usage{
.apply_aggregator(aggregator, data)
}
\arguments{
\item{aggregator}{Object of class aggregator.}

\item{data}{Vector or matrix to be aggregates}
}
\value{
Aggregated data
}
\description{
Internal helper function to apply aggregator to data
}
