% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{.apply_mask}
\alias{.apply_mask}
\title{Apply mask to extract the elements of the matrix that are both true}
\usage{
.apply_mask(mat, indices_1, indices_2)
}
\arguments{
\item{mat}{Matrix containing elements for extraction.}

\item{indices_1}{Matrix of booleans of the same dimensions of \code{mat}.}

\item{indices_2}{Matrix of booleans of the same dimensions of \code{mat}}
}
\value{
Matrix of same dimensions of \code{mat} with the overlapping \code{TRUE}
elements only.
}
\description{
Apply mask to extract the elements of the matrix that are both true
}
\keyword{internal}
