% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_reference_times.R
\name{.assign_allocation_from_ns}
\alias{.assign_allocation_from_ns}
\title{Assign number of reference times to delay and uncertainty from the sizes}
\usage{
.assign_allocation_from_ns(
  n_used,
  n_target,
  n_min_delay,
  n_min_retro_nowcasts,
  n_required,
  prop_delay
)
}
\arguments{
\item{n_used}{Integer indicating number reference times that will be used.}

\item{n_target}{Integer indicating the target number of reference times.}

\item{n_min_delay}{Integer indicating the number needed for delay
estimation.}

\item{n_min_retro_nowcasts}{Integer indicating the number needed for
uncertainty estimation.}

\item{n_required}{Integer indicating the number need for both delay and
uncertainty}

\item{prop_delay}{Numeric value <1 indicating what proportion of all
reference times in the reporting triangle to be used for delay
estimation. Default is \code{0.5}.}
}
\value{
List of number of reference times to use for delay and uncertainty
}
\description{
Assign number of reference times to delay and uncertainty from the sizes
}
