% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_reference_times.R
\name{.assign_ns_from_sizes}
\alias{.assign_ns_from_sizes}
\title{Assign number of reference times to delay and uncertainty from the sizes}
\usage{
.assign_ns_from_sizes(
  size_used,
  size_target,
  size_min_delay,
  size_min_retro_nowcasts,
  size_required,
  prop_delay
)
}
\arguments{
\item{size_used}{Integer indicating number reference times that will be used.}

\item{size_target}{Integer indicating the target number of reference times.}

\item{size_min_delay}{Integer indicating the number needed for delay
estimation.}

\item{size_min_retro_nowcasts}{Integer indicating the number needed for
uncertainty estimation.}

\item{size_required}{Integer indicating the number need for both delay and
uncertainty}

\item{prop_delay}{Numeric value <1 indicating what proportion of all
reference times in the reporting triangle to be used for delay
estimation. Default is \code{0.5}.}
}
\value{
List of number of reference times to use for delay and uncertainty
}
\description{
Assign number of reference times to delay and uncertainty from the sizes
}
