% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.autoescape_brackets}
\alias{.autoescape_brackets}
\title{Escape brackets returned in a string for passing to glue}
\usage{
.autoescape_brackets(string)
}
\arguments{
\item{string}{Vector of strings containing \code{{}}}
}
\value{
Vector of strings where all single brackets are replaced with double
brackets
}
\description{
Escape brackets returned in a string for passing to glue
}
\keyword{internal}
