% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_reference_times.R
\name{.calculate_ns}
\alias{.calculate_ns}
\title{Helper function to calculate various size requirements}
\usage{
.calculate_ns(
  reporting_triangle,
  max_delay,
  scale_factor,
  n_min_retro_nowcasts
)
}
\arguments{
\item{reporting_triangle}{A \link{reporting_triangle} object with rows
representing reference times and columns representing delays.
Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row (e.g., weekly reporting of daily data).}

\item{scale_factor}{Numeric value indicating the multiplicative factor on
the maximum delay to be used for estimation of delay and uncertainty.
Default is \code{3}.}

\item{n_min_retro_nowcasts}{Integer indicating the minimum number of
reference times needed for uncertainty estimation. Default is \code{2}.}
}
\value{
list of the integer sizes
}
\description{
Helper function to calculate various size requirements
}
\keyword{internal}
