% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_reference_times.R
\name{.check_against_requirements}
\alias{.check_against_requirements}
\title{Check target size against number of reference times available and the number
required}
\usage{
.check_against_requirements(n_ref_times, size_required, size_target)
}
\arguments{
\item{n_ref_times}{Integer indicating the number of reference tiems available}

\item{size_required}{Integer indicating the number need for both delay and
uncertainty}

\item{size_target}{Integer indicating the target number of reference times.}
}
\value{
\code{size_used} Integer indicating how many reference times will be used
}
\description{
Check target size against number of reference times available and the number
required
}
