% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_lhs_for_zeros}
\alias{.check_lhs_for_zeros}
\title{Check that there are not only 0s on the LHS of NAs}
\usage{
.check_lhs_for_zeros(mat)
}
\arguments{
\item{mat}{Matrix to check}
}
\value{
Boolean indicating whether or not there are only 0s on the LHS
of the first NA
}
\description{
Check that there are not only 0s on the LHS of NAs
}
