% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.check_obs_and_pred}
\alias{.check_obs_and_pred}
\title{Check observations and predictions are compatible}
\usage{
.check_obs_and_pred(obs, pred)
}
\arguments{
\item{obs}{Matrix or vector of observations.}

\item{pred}{Matrix or vector of predictions.}
}
\value{
NULL, invisibly
}
\description{
Check observations and predictions are compatible
}
\keyword{internal}
