% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.check_to_filter_to_max_delay}
\alias{.check_to_filter_to_max_delay}
\title{Check that the reporting triangle contains the correct number of columns for
the specified maximum delay}
\usage{
.check_to_filter_to_max_delay(triangle, max_delay)
}
\arguments{
\item{max_delay}{Integer indicating the maximum delay to estimate, in units
of the delay. The default is to use the whole reporting triangle,
\code{ncol(reporting_triangle) -1}.}
}
\value{
reporting_triangle
}
\description{
Check that the reporting triangle contains the correct number of columns for
the specified maximum delay
}
