% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_zeros_bottom_right}
\alias{.check_zeros_bottom_right}
\title{Check if matrix only contains zeros in the bottom right}
\usage{
.check_zeros_bottom_right(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Boolean indicating whether the matrix only contains zeros in the
bottom right (if TRUE, entire bottom right is 0s)
}
\description{
Check if matrix only contains zeros in the bottom right
}
\keyword{internal}
