% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{.combine_data}
\alias{.combine_data}
\title{Combine data from a nowcast dataframe, strata, and reference dates}
\usage{
.combine_data(nowcast_df, strata_list, reference_dates)
}
\arguments{
\item{nowcast_df}{Data.frame containing information for multiple draws with
columns for the reference time (\code{time}), the predicted counts
(\code{pred_count}), and the draw number (\code{draw}).}

\item{strata_list}{Named list where each entry should be the name of the strata
and the name of the corresponding strata.}

\item{reference_dates}{Vector of reference dates corresponding to the
reference times in the \code{nowcast_df}.}
}
\value{
Data.frame indexed by the reference dates and with columns for
each of the named elements in the named list \code{strata}.
}
\description{
Combines data from a nowcast dataframe, a named list of the
strata associated with the nowcast dataframe, and a vector of reference
dates corresponding to the time column in the \code{nowcast_df}
}
