% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-methods.R
\name{.compute_nowcast_stats}
\alias{.compute_nowcast_stats}
\title{Compute nowcast requirement statistics}
\usage{
.compute_nowcast_stats(mat)
}
\arguments{
\item{mat}{Matrix representation of the triangle.}
}
\value{
List with dates_need_nowcast and dates_complete.
}
\description{
Compute nowcast requirement statistics
}
\keyword{internal}
