% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{.compute_zero_stats}
\alias{.compute_zero_stats}
\title{Compute zero value statistics}
\usage{
.compute_zero_stats(mat)
}
\arguments{
\item{mat}{Matrix representation of the triangle.}
}
\value{
List with num_zeros, pct_zeros, and zeros_per_row.
}
\description{
Compute zero value statistics
}
\keyword{internal}
