% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{.conditional_sum_cols}
\alias{.conditional_sum_cols}
\title{Compute the sum of entries of a column in a matrix where both sets of
matrices of booleans are TRUE}
\usage{
.conditional_sum_cols(col, matrix_bool1, matrix_bool2, matrix_to_sum)
}
\arguments{
\item{col}{Integer indicating the column to sum over}

\item{matrix_bool1}{Matrix of booleans for the first set of indices}

\item{matrix_bool2}{Matrix of booleans for the second set of indices}

\item{matrix_to_sum}{Matrix to be summed for that specific column}
}
\value{
Numeric summing the values in the \code{matrix_to_sum} at the specified
column for the entries that are true
}
\description{
Compute the sum of entries of a column in a matrix where both sets of
matrices of booleans are TRUE
}
\keyword{internal}
