% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{.count_negative_rows}
\alias{.count_negative_rows}
\title{Count rows with negative values}
\usage{
.count_negative_rows(mat)
}
\arguments{
\item{mat}{Matrix representation of the triangle.}
}
\value{
Integer count of rows with at least one negative value.
}
\description{
Count rows with negative values
}
\keyword{internal}
