% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-getters.R
\name{.diff_days}
\alias{.diff_days}
\title{Internal: Compute day difference between dates}
\usage{
.diff_days(report_date, reference_date)
}
\arguments{
\item{report_date}{Date vector of report dates}

\item{reference_date}{Date vector of reference dates}
}
\value{
Numeric vector of delays
}
\description{
Internal: Compute day difference between dates
}
\keyword{internal}
