% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_function_name}
\alias{.extract_function_name}
\title{Extract Base Function Name
Adopted from https://github.com/epinowcast/primarycensored/blob/04f3ff2e5c8741c1f6e003566c68bbaf491cb073/R/utils.R#L11 #nolint
This helper function extracts the base name of a function, removing an
namespace prefixes.}
\usage{
.extract_function_name(func)
}
\arguments{
\item{func}{The output of \code{substitute} on a function.}
}
\value{
A character string representing the base name of the function.
}
\description{
Extract Base Function Name
Adopted from https://github.com/epinowcast/primarycensored/blob/04f3ff2e5c8741c1f6e003566c68bbaf491cb073/R/utils.R#L11 #nolint
This helper function extracts the base name of a function, removing an
namespace prefixes.
}
\keyword{internal}
