% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_predictions}
\alias{.extract_predictions}
\title{Extract from one matrix only elements that are missing in another}
\usage{
.extract_predictions(point_nowcast_matrix, reporting_triangle)
}
\value{
Matrix containing the elements from \code{point_nowcast_matrix} for
only the elements that are missing in \code{reporting_triangle}
}
\description{
Extract from one matrix only elements that are missing in another
}
\keyword{internal}
