% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{.filter_to_recent_horizons}
\alias{.filter_to_recent_horizons}
\title{Filter to recent horizons}
\usage{
.filter_to_recent_horizons(mat, n_possible_horizons)
}
\arguments{
\item{mat}{Matrix containing all the rows}

\item{n_possible_horizons}{Number of rows we want starting from the final
row}
}
\value{
\code{bottom_matrix} Matrix containing the last \code{n_possible_horizons}
rows of the matrix.
}
\description{
Filter to recent horizons
}
