% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_obs_vs_pred.R
\name{.fit_distrib}
\alias{.fit_distrib}
\title{Fit observations vs predictions to an observation model}
\usage{
.fit_distrib(obs, pred, observation_model = "dnbinom")
}
\arguments{
\item{obs}{Matrix of observations with retrospective nowcast dates as rows
and horizons as columns.}

\item{pred}{Matrix of predictions with retrospective nowcast dates as rows
and horizons as columns.}

\item{observation_model}{Character string indicating the choice of
observation model to fit to the predicted nowcasts versus the
observations. Default is \verb{negative binomial}.}
}
\value{
Vector of parameters corresponding to the chosen
\code{observation_model} of length of the number of columns in \code{obs} and
\code{pred}, with each element corresponding to a horizon.
disp <- .fit_distrib(obs, pred, observation_model = "negative binomial")
}
\description{
Fit observations vs predictions to an observation model
}
