% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{.get_delay_nowcast_and_obs_t}
\alias{.get_delay_nowcast_and_obs_t}
\title{Get expected and observed additional counts as of reference time t}
\usage{
.get_delay_nowcast_and_obs_t(t, matr_observed, n_history)
}
\arguments{
\item{t}{Integer indicating the "as of" time}

\item{matr_observed}{Matrix of the available observations, with NAs for
the observations that are missing.}

\item{n_history}{Integer indicating the number of observations to be
used in the estimate of the delay estimate for each retrospective reporting
triangle, if the delay estimate is being re-estimated.}
}
\value{
Dataframe of length of \code{delay_pmf} containing the expected values
to be added at each delay as of time t and the observed values to be added
as of time t
}
\description{
Get expected and observed additional counts as of reference time t
}
\keyword{internal}
