% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{.get_exp_and_obs_as_of_t}
\alias{.get_exp_and_obs_as_of_t}
\title{Get expected and observed additional counts as of reference time t}
\usage{
.get_exp_and_obs_as_of_t(t, matr_observed, delay_pmf)
}
\arguments{
\item{t}{Integer indicating the "as of" time}

\item{matr_observed}{Matrix of the available observations, with NAs for
the observations that are missing.}

\item{delay_pmf}{Vector of delays assumed to be indexed starting at the
first delay column in \code{matr_observed}}
}
\value{
Dataframe of length of \code{delay_pmf} containing the expected values
to be added at each delay as of time t and the observed values to be added
as of time t
}
\description{
Get expected and observed additional counts as of reference time t
}
\keyword{internal}
