% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{.get_triangle_info}
\alias{.get_triangle_info}
\title{Get triangle dimension information}
\usage{
.get_triangle_info(x)
}
\arguments{
\item{x}{A \link{reporting_triangle} object.}
}
\value{
List with components: n_rows, n_cols, max_delay, delays_unit,
structure.
}
\description{
Internal helper to extract key dimension information from a triangle.
}
\keyword{internal}
