% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_reference_times.R
\name{.handle_target_exceeds_avail}
\alias{.handle_target_exceeds_avail}
\title{Helper for when target exceeds available reference times}
\usage{
.handle_target_exceeds_avail(
  n_ref_times,
  size_required,
  size_target,
  size_min_delay,
  size_min_retro_nowcasts
)
}
\arguments{
\item{n_ref_times}{Integer indicating the number of reference times
available}

\item{size_required}{Integer indicating the number need for both delay and
uncertainty}

\item{size_target}{Integer indicating the target number of reference times.}

\item{size_min_delay}{Integer indicating the number needed for delay
estimation.}

\item{size_min_retro_nowcasts}{Integer indicating the number needed for
uncertainty estimation.}
}
\value{
number of reference times to use or NULL, invisibly
}
\description{
Helper for when target exceeds available reference times
}
