% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.is_date_convertible}
\alias{.is_date_convertible}
\title{Test if something can be converted to a date object using as.Date}
\usage{
.is_date_convertible(x)
}
\arguments{
\item{x}{Single or vector of either a character string or Date class to
be tested.}
}
\value{
Boolean indicating whether x can be converted to a Date object
}
\description{
Test if something can be converted to a date object using as.Date
}
