% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_reporting_triangle}
\alias{.is_reporting_triangle}
\title{Check if matrix is a reporting triangle}
\usage{
.is_reporting_triangle(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Boolean indicating whether the matrix contains only NAs in the
bottom right and not NAs everywhere else
}
\description{
Check if matrix is a reporting triangle
}
\keyword{internal}
