% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{.nowcast_from_rep_tris}
\alias{.nowcast_from_rep_tris}
\title{Produce a nowcast from a named reporting triangle object}
\usage{
.nowcast_from_rep_tris(
  rep_tri,
  name,
  strata_cols,
  scale_factor,
  prop_delay,
  output_type,
  draws,
  uncertainty_model,
  uncertainty_sampler,
  delay_pmf,
  uncertainty_params,
  ref_time_aggregator = identity,
  delay_aggregator = function(x) rowSums(x, na.rm = TRUE)
)
}
\arguments{
\item{rep_tri}{Object of reporting triangle class}

\item{name}{Name of the reporting triangle class object}

\item{strata_cols}{Vector of character strings indicating the names of the
columns in \code{data} that determine how to stratify the data for nowcasting.
The unique combinations of the entries in the \code{strata_cols} denote the
unit of a single nowcast. Within a strata, there can be no repeated
unique combinations of reference dates and report dates. Default is \code{NULL}
which assumes that the data.frame being passed in represents a single
strata (only one nowcast will be produced). All columns that are not
part of the \code{strata_cols} will be removed.}

\item{scale_factor}{Numeric value indicating the multiplicative factor on
the maximum delay to be used for estimation of delay and uncertainty.
Default is \code{3}.}

\item{prop_delay}{Numeric value <1 indicating what proportion of all
reference times in the reporting triangle to be used for delay
estimation. Default is \code{0.5}.}

\item{output_type}{Character string indicating whether the output should be
samples (\code{"samples"})from the estimate with full uncertainty or whether to
return the point estimate (\code{"point"}). Default is \code{"samples"}.}

\item{draws}{Integer indicating the number of probabilistic draws to include
if \code{output_type} is \code{"samples"}. Default is 1000.}

\item{uncertainty_model}{Function that ingests a matrix of observations and a
matrix of predictions and returns a vector that can be used to
apply uncertainty using the same error model. Default is
\code{fit_by_horizon} with arguments of \code{obs} matrix of observations and
\code{pred} the matrix of predictions that fits each column (horizon)
to a negative binomial observation model by default. The user can
specify a different fitting model by replacing the
\code{fit_model} argument in \code{fit_by_horizon}.}

\item{uncertainty_sampler}{Function that ingests a vector or matrix of
predictions and a vector of uncertainty parameters and generates draws
from the observation model. Default is \code{sample_nb} which expects
arguments \code{pred} for the vector of predictions and uncertainty parameters
for the corresponding vector of uncertainty parameters, and draws from a
negative binomial for each element of the vector.}

\item{delay_pmf}{Vector of delays assumed to be indexed starting at the
first delay column in \code{data$reporting_triangle_matrix}. Default is NULL,
which will estimate the delay from the reporting triangle matrix in \code{data},
See \code{\link{estimate_delay}} for more details.}

\item{uncertainty_params}{Vector of uncertainty parameters ordered from
horizon 1 to the maximum horizon. Default is \code{NULL}, which will
result in computing the uncertainty parameters from the reporting
triangle matrix \code{data}. See \code{\link{estimate_uncertainty}} for more
details.}

\item{ref_time_aggregator}{Function that operates along the rows (reference
times) of the retrospective point nowcast matrix before it has been
aggregated across columns (delays). Default is \code{identity}
which does not aggregate across reference times.}

\item{delay_aggregator}{Function that operates along the columns (delays)
of the retrospective point nowcast matrix after it has been aggregated
across reference times. Default is \code{function(x) rowSums(x, na.rm = TRUE)}.}
}
\value{
Data.frame of nowcasts with columns obtained from the name of the
original list and the nowcast unit.
}
\description{
Produce a nowcast from a named reporting triangle object
}
\keyword{internal}
