% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{.rename_cols}
\alias{.rename_cols}
\title{Rename required columns}
\usage{
.rename_cols(data, old_names)
}
\arguments{
\item{data}{Data.frame with the original column names}

\item{old_names}{The names of the columns for, in order, reference date,
report date, and count}
}
\value{
Data.frame with columns renamed
}
\description{
Rename required columns
}
