% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.return_to_name_cols}
\alias{.return_to_name_cols}
\title{Rename required columns back to original names}
\usage{
.return_to_name_cols(data, old_names)
}
\arguments{
\item{data}{Data.frame with the standard column names}

\item{old_names}{The original names of the columns for, in order,
reference date, report date, and count}
}
\value{
Data.frame with columns renamed back to their original values
}
\description{
Rename required columns back to original names
}
