% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{.split_df_by_cols}
\alias{.split_df_by_cols}
\title{Split dataframe into a list of dataframes by the entries in the specified
columns}
\usage{
.split_df_by_cols(long_df, col_names)
}
\arguments{
\item{long_df}{Data.frame to be split into a list of dataframes.}

\item{col_names}{Character string indicating the column names to be used
to create the new data.frames.}
}
\value{
List of data.frames named by the concatenated entries in col_names
}
\description{
Split dataframe into a list of dataframes by the entries in the specified
columns
}
\keyword{internal}
