% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_aggregation_function}
\alias{.validate_aggregation_function}
\title{Validate aggregation function
Checks that the aggregation function is one of the allowed functions.}
\usage{
.validate_aggregation_function(fun_to_aggregate)
}
\arguments{
\item{fun_to_aggregate}{Function to validate}
}
\value{
NULL, invisibly
}
\description{
Validate aggregation function
Checks that the aggregation function is one of the allowed functions.
}
\keyword{internal}
