% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_date_cols}
\alias{.validate_date_cols}
\title{Validate the date columns can be converted to Dates}
\usage{
.validate_date_cols(data, date_col)
}
\arguments{
\item{data}{Data.frame containing the date_col}

\item{date_col}{Name of the column to be tested}
}
\description{
Validate the date columns can be converted to Dates
}
\keyword{internal}
