% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_delay}
\alias{.validate_delay}
\title{Validate the delay PMF if it is passed in}
\usage{
.validate_delay(triangle, delay_pmf)
}
\arguments{
\item{triangle}{Matrix of values with rows indicating the time points and
columns indicating the delays.}

\item{delay_pmf}{Vector of length of the number of delays indicating the
probability of a case being reported on a given delay.}
}
\value{
NULL invisibly
}
\description{
Validate the delay PMF if it is passed in
}
\keyword{internal}
