% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_delay_and_triangle}
\alias{.validate_delay_and_triangle}
\title{Validate triangle to nowcast and delay pmf together}
\usage{
.validate_delay_and_triangle(triangle, delay_pmf)
}
\arguments{
\item{triangle}{Matrix of values with rows indicating the time points and
columns indicating the delays}

\item{delay_pmf}{Vector of length of the number of delays indicating the
probability of a case being reportined on a given delay}
}
\value{
NULL, invisibly
}
\description{
Various checks to make sure that the reporting triangle  and the delay pmf
passed in to \code{\link[=apply_delay]{apply_delay()}} are formatted properly and compaitble
}
\keyword{internal}
