% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_delays_unit}
\alias{.validate_delays_unit}
\title{Validate delays unit parameter}
\usage{
.validate_delays_unit(delays_unit)
}
\arguments{
\item{delays_unit}{Character string indicating the unit of delays.}
}
\value{
NULL invisibly
}
\description{
Validate delays unit parameter
}
\keyword{internal}
