% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_fit_obs_model}
\alias{.validate_fit_obs_model}
\title{Validate observation model name}
\usage{
.validate_fit_obs_model(observation_model_name)
}
\arguments{
\item{observation_model_name}{Character string indicating the choice of
observation model to fit to the predicted nowcasts versus the
observations. Default is \code{"negative binomial"}.}
}
\value{
NULL or error message
}
\description{
Validate observation model name
}
