% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_for_delay_estimation}
\alias{.validate_for_delay_estimation}
\title{Validate reporting_triangle for delay estimation
Domain-specific checks to ensure the reporting triangle is suitable for
delay estimation in \code{\link[=estimate_delay]{estimate_delay()}}.}
\usage{
.validate_for_delay_estimation(triangle, n = nrow(triangle))
}
\arguments{
\item{triangle}{Matrix of values with rows indicating the time points and
columns indicating the delays.}

\item{n}{Integer indicating the number of reference times (observations) to
be used in the estimate of the reporting delay, always starting from the
most recent reporting delay. The default is to use the whole reporting
triangle, so \code{nrow(reporting_triangle)}.}
}
\value{
NULL, invisibly
}
\description{
Validate reporting_triangle for delay estimation
Domain-specific checks to ensure the reporting triangle is suitable for
delay estimation in \code{\link[=estimate_delay]{estimate_delay()}}.
}
\keyword{internal}
