% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_inputs_allocation}
\alias{.validate_inputs_allocation}
\title{Helper function to validate allocation parameters}
\usage{
.validate_inputs_allocation(scale_factor, prop_delay, n_min_retro_nowcasts)
}
\arguments{
\item{scale_factor}{Numeric value indicating the multiplicative factor on
the maximum delay to be used for estimation of delay and uncertainty.
Default is \code{3}.}

\item{prop_delay}{Numeric value <1 indicating what proportion of all
reference times in the reporting triangle to be used for delay
estimation. Default is \code{0.5}.}

\item{n_min_retro_nowcasts}{Integer indicating the minimum number of
reference times needed for uncertainty estimation. Default is \code{2}.}
}
\value{
NULL invisibly
}
\description{
Helper function to validate allocation parameters
}
