% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_max_delay}
\alias{.validate_max_delay}
\title{Check that the maximum delay is not too large, error if it is}
\usage{
.validate_max_delay(reporting_triangle, max_delay)
}
\arguments{
\item{max_delay}{Integer indicating the maximum delay to estimate, in units
of the delay. The default is to use the whole reporting triangle,
\code{ncol(reporting_triangle) -1}.}
}
\value{
NULL invisibly
}
\description{
Check that the maximum delay is not too large, error if it is
}
