% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_multiple_inputs}
\alias{.validate_multiple_inputs}
\title{Validate the inputs to \code{estimate_and_apply_uncertainty()} to ensure that
the reporting triangle and point nowcast matrix are compatible.}
\usage{
.validate_multiple_inputs(point_nowcast_matrix, reporting_triangle)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of point nowcast predictions and
observations, with rows representing the reference times and columns
representing the delays.}

\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}
}
\value{
NULL invisibly
}
\description{
Validate the inputs to \code{estimate_and_apply_uncertainty()} to ensure that
the reporting triangle and point nowcast matrix are compatible.
}
\keyword{internal}
