% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_nowcast_unit}
\alias{.validate_nowcast_unit}
\title{Validate each the nowcast unit passed to baselinenowcast}
\usage{
.validate_nowcast_unit(nowcast_unit, data, reference_date, report_date, count)
}
\arguments{
\item{nowcast_unit}{Vector of character strings indicating the names of the
columns in \code{data} that denote the unit of a single nowcast. Within a
nowcast unit, there can be no repeated unique combinations of reference
dates and report dates. Default is \code{NULL} which assumes that all columns
that are not required columns (reference date, report date, count) form
the unit of a single forecast. This may lead to unexpected behaviour, so
setting the nowcast unit explicitly can help make the code easier to debug
and easier to read. If specified, all columns that are not part of the
forecast unit (or required columns) will be removed.}

\item{data}{Data.frame in a long tidy format with counts by reference date
and report date for one or more strata. Must contain the following
columns:
.    - Column of type \code{date} or character with the dates of
the primary event occurrence (reference date).
\itemize{
\item Column of type \code{date} or character with the dates of
report of the primary event (report_date).
\item Column of numeric or integer indicating the new confirmed counts
pertaining to that reference and report date (count).
Additional columns can be included, and the user can specify which columns
set the unit of a single nowcast  ( i.e. the combination of columns that
uniquely define a single nowcast with the \code{nowcast_unit} argument.).
}}

\item{reference_date}{Character string indicating the name of the
column which represents the reference date, or the date of the primary
event occurrence.}

\item{report_date}{Character string indicating the name of the
column which represents the date the primary event was reported.}

\item{count}{Character string indicating the name of the column
containing the number of incident cases on each reference and report date.}
}
\description{
Validate each the nowcast unit passed to baselinenowcast
}
