% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_ref_time_allocation}
\alias{.validate_ref_time_allocation}
\title{Validate the specified number of reference times meets the minimum
requirements}
\usage{
.validate_ref_time_allocation(
  n_ref_times,
  size_min_ref_times_delay,
  n_history_delay,
  n_retrospective_nowcasts,
  size_min_retro_nowcasts = 2
)
}
\arguments{
\item{n_ref_times}{Integer indicating the number of reference times
available.}

\item{size_min_ref_times_delay}{Integer indicating minimum number of
reference times needed for delay estimation.}

\item{n_history_delay}{Integer indicating the number of reference times
(observations) to be used in the estimate of the reporting delay, always
starting from the most recent reporting delay.}

\item{n_retrospective_nowcasts}{Integer indicating the number of
retrospective nowcast times to use for uncertainty estimation.}
}
\value{
NULL, invisibly
}
\description{
Validate the specified number of reference times meets the minimum
requirements
}
