% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_rep_tri_args}
\alias{.validate_rep_tri_args}
\title{Validate each item in the reporting triangle}
\usage{
.validate_rep_tri_args(reporting_triangle_matrix, reference_dates, delays_unit)
}
\arguments{
\item{reporting_triangle_matrix}{Matrix of reporting triangle where rows
are reference times, columns are delays, and entries are incident counts.}

\item{reference_dates}{Vector of Date objects indicating the reference dates
corresponding to each row of the matrix.}

\item{delays_unit}{Character string specifying the temporal granularity.
Options are \code{"days"}, \code{"weeks"}, \code{"months"}, \code{"years"}.}
}
\description{
Validate each item in the reporting triangle
}
\keyword{internal}
