% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_rep_tri_df}
\alias{.validate_rep_tri_df}
\title{Validate the reporting triangle data.frame}
\usage{
.validate_rep_tri_df(data)
}
\arguments{
\item{data}{Data.frame in long tidy form with reference dates, report dates,
and case counts, used to create a \code{reporting_triangle} object.}
}
\value{
NULL, invisibly
}
\description{
Validate the reporting triangle data.frame
}
