% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_retro_inputs}
\alias{.validate_retro_inputs}
\title{Validate inputs use to generate retrospective data
Various checks to make sure that the reporting triangle and specified
integers passed in to \code{\link[=generate_retrospective_data]{generate_retrospective_data()}} are formatted
properly}
\usage{
.validate_retro_inputs(triangle, n_history_uncertainty, n_history_delay)
}
\arguments{
\item{n_history_uncertainty}{Integer indicating the number of retrospective
reporting triangles to be generated, always starting from the most
recent reference time. The default is to generate as many retrospective
reporting triangles that would have a sufficient number of rows to
estimate  a delay distribution and generate a nowcast from.}

\item{n_history_delay}{Integer indicating the number of reference dates to be
used in the estimate of the reporting delay, always starting from the most
recent reporting delay. The default here is to us the difference between
the number of observation (rows) in \code{triangle_for_retro} and
\code{n_history_uncertainty} so all of the data is being used.}
}
\value{
NULL, invisibly
}
\description{
Validate inputs use to generate retrospective data
Various checks to make sure that the reporting triangle and specified
integers passed in to \code{\link[=generate_retrospective_data]{generate_retrospective_data()}} are formatted
properly
}
\keyword{internal}
