% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_uncertainty}
\alias{.validate_uncertainty}
\title{Validate the uncertainty parameters if they are passed in}
\usage{
.validate_uncertainty(triangle, uncertainty_params)
}
\arguments{
\item{triangle}{Matrix of values with rows indicating the time points and
columns indicating the delays.}

\item{uncertainty_params}{Vector of uncertainty parameters ordered from
horizon 1 to the maximum horizon. Note that these will be reversed
internally to match the ordering of the \code{point_nowcast_matrix} (where
a horizon of 1 is the last entry).}
}
\value{
NULL invisibly
}
\description{
Validate the uncertainty parameters if they are passed in
}
