% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_uncertainty_inputs}
\alias{.validate_uncertainty_inputs}
\title{Validate the triangle used to estimate uncertainty and the specified
inputs together, as only certain combinations are valid}
\usage{
.validate_uncertainty_inputs(
  triangle_for_uncertainty,
  delay_pmf,
  n_history_dispersion,
  n_history
)
}
\arguments{
\item{triangle_for_uncertainty}{Matrix of values with rows indicating the
time points and columns indicating the delays. Specifically, the triangle
used to generate the uncertainty estimates.}

\item{delay_pmf}{Vector of length of the number of delays indicating the
probability of a case being reported on a given delay}

\item{n_history_dispersion}{Integer indicating the number of observations
to be used to estimate the dispersion parameters in the observation model}

\item{n_history}{Integer indicating the number of observations to be used to
estimate the delay distribution in order to generate retrospective
nowcasts}
}
\value{
NULL, invisibly
}
\description{
Validate the triangle used to estimate uncertainty and the specified
inputs together, as only certain combinations are valid
}
\keyword{internal}
