% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{.validate_uncertainty_matrices}
\alias{.validate_uncertainty_matrices}
\title{Validate lists of matrices for uncertainty estimation}
\usage{
.validate_uncertainty_matrices(list_of_ncs, list_of_obs)
}
\arguments{
\item{list_of_ncs}{List of nowcast matrices}

\item{list_of_obs}{List of observation matrices}
}
\value{
NULL (invisibly) if valid, otherwise throws an error
}
\description{
Validate lists of matrices for uncertainty estimation
}
\keyword{internal}
