% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_and_apply_delay.R
\name{estimate_and_apply_delay}
\alias{estimate_and_apply_delay}
\title{Estimate and apply delay from a reporting triangle}
\usage{
estimate_and_apply_delay(reporting_triangle, ...)
}
\arguments{
\item{reporting_triangle}{A \link{reporting_triangle} object with rows
representing reference times and columns representing delays.
Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row (e.g., weekly reporting of daily data).}

\item{...}{Additional arguments passed to the \code{estimate_delay} function}
}
\value{
\code{pt_nowcast_matrix} A \code{reporting_triangle} object of point nowcasts
with the same structure as the input
}
\description{
Estimate and apply delay from a reporting triangle
}
\examples{
triangle_mat <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, NA,
    80, 40, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 5,
  byrow = TRUE
)
ref_dates <- seq(
  from = as.Date("2025-01-01"),
  by = "day",
  length.out = nrow(triangle_mat)
)
triangle <- as_reporting_triangle(
  data = triangle_mat,
  reference_dates = ref_dates,
  max_delay = 3
)
pt_nowcast_matrix <- estimate_and_apply_delay(
  reporting_triangle = triangle,
  n = 4
)
pt_nowcast_matrix
}
\seealso{
High-level workflow wrapper functions
\code{\link{allocate_reference_times}()},
\code{\link{estimate_and_apply_uncertainty}()},
\code{\link{estimate_uncertainty_retro}()}
}
\concept{workflow_wrappers}
