% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_and_apply_uncertainty.R
\name{estimate_and_apply_uncertainty}
\alias{estimate_and_apply_uncertainty}
\title{Estimate and apply uncertainty to a point nowcast matrix}
\usage{
estimate_and_apply_uncertainty(
  point_nowcast_matrix,
  reporting_triangle,
  max_delay,
  n_history_delay = round(1.5 * max_delay),
  n_retrospective_nowcasts = 3 * max_delay - n_history_delay,
  draws = 100,
  uncertainty_model = fit_by_horizon,
  uncertainty_sampler = sample_nb,
  ref_time_aggregator = identity,
  delay_aggregator = function(x) rowSums(x, na.rm = TRUE),
  ...
)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of point nowcast predictions and
observations, with rows representing the reference times and columns
representing the delays.}

\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{max_delay}{Integer indicating the maximum delay to estimate, in units
of the delay. The default is to use the whole reporting triangle,
\code{ncol(reporting_triangle) -1}.}

\item{n_retrospective_nowcasts}{Integer indicating the number of reference
times to use as retrospective nowcast time. Default is 1.5* the maximum
delay}

\item{draws}{Integer indicating the number of draws of the predicted
nowcast vector to generate. Default is \code{1000}.}

\item{...}{Additional arguments}

\item{error_model}{Character string indicating what error model to use.
Default is \code{"negative binomial"}}

\item{aggregator}{Function to aggregate predictions for error estimation.
Default is \code{sum}.}
}
\value{
\code{nowcast_draws_df} Dataframe containing draws of combined
observations and probabilistic predictions at each reference time.
}
\description{
Estimate and apply uncertainty to a point nowcast matrix
}
\examples{
triangle <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, NA,
    80, 40, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 5,
  byrow = TRUE
)
pt_nowcast_matrix <- estimate_and_apply_delay(
  reporting_triangle = triangle,
  max_delay = 3,
  n = 4
)
nowcast_draws_df <- estimate_and_apply_uncertainty(
  pt_nowcast_matrix,
  triangle,
  max_delay,
  n_retrospective_nowcasts = 1
)
nowcast_draws_df
}
