% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_disp_params.R
\name{estimate_disp_params}
\alias{estimate_disp_params}
\title{Estimate dispersion parameters}
\usage{
estimate_disp_params(
  list_of_nowcasts,
  latest_rep_tri,
  n = length(list_of_nowcasts)
)
}
\arguments{
\item{list_of_nowcasts}{List of complete reporting rectangles where rows
represent reference time points and columns represent delays.}

\item{latest_rep_tri}{Matrix of the latest, incomplete, reporting triangle,
with rows representing the time points of reference and columns
representing the delays.}

\item{n}{Integer indicating the number of reporting rectangles to use to
estimate the dispersion parameters}
}
\value{
Vector of length one less than the number of columns in the
latest reporting triangle, with each element representing the estimate
of the dispersion parameter for each delay d, starting at delay d=1.
}
\description{
This function ingests a list of nowcasts (completed reporting rectangles) and
the latest reporting triangle and uses both to estimate dispersion
parameters from the observations and estimates at each delay.
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

trunc_rts <- truncate_triangles(
  triangle = triangle,
  n = 2
)
retro_rts <- generate_triangles(
  truncated_triangles = trunc_rts
)
retro_nowcasts <- generate_nowcasts(
  list_of_rts = retro_rts,
  n = 5
)
disp_params <- estimate_disp_params(
  list_of_nowcasts = retro_nowcasts,
  latest_rep_tri = triangle
)
}
