% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_missing_elements}
\alias{extract_missing_elements}
\title{Extract from one matrix only elements that are missing in another}
\usage{
extract_missing_elements(full_mat, subset_mat)
}
\arguments{
\item{full_mat}{Matrix containing values at all elements}

\item{subset_mat}{Matrix containing missing values for some elements}
}
\value{
\code{other_subset_mat} Matrix containing the elements from \code{full_mat}
for only the elements that are missing in \code{subset_mat}
}
\description{
Extract from one matrix only elements that are missing in another
}
\examples{
full_mat <- matrix(
  c(
    1, 3, 5, 7,
    4, 7, 8, 9,
    9, 10, 3, 5,
    3, 4, 8, 5
  ),
  nrow = 4,
  byrow = TRUE
)

subset_mat <-matrix(
  c(
    1, 3, 5, 7,
    4, 7, 8, NA,
    9, 10, NA, NA,
    3, NA, NA, NA
  ),
  nrow = 4,
  byrow = TRUE
)

other_subset <- extract_missing_elements(full_mat, subset_mat)
other_subset
}
