% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_observed_nowcast_components.R
\name{extract_obs_and_pred}
\alias{extract_obs_and_pred}
\title{Extract observations and predictions by retrospective nowcast date and
horizon}
\usage{
extract_obs_and_pred(
  pt_nowcast_matrices,
  trunc_reporting_triangles,
  retro_reporting_triangles,
  n = length(pt_nowcast_matrices),
  fun_to_aggregate = sum,
  k = 1
)
}
\arguments{
\item{pt_nowcast_matrices}{List of point nowcast matrices where rows
represent reference time points and columns represent delays.}

\item{trunc_reporting_triangles}{List of truncated reporting matrices,
containing all observations as of the latest reference time. Elements of
list are paired with elements of \code{pt_nowcast_matrices}.}

\item{retro_reporting_triangles}{List of \code{n} truncated reporting triangle
matrices with as many rows as available given the truncation.}

\item{n}{Integer indicating the number of reporting matrices to use to
estimate the dispersion parameters.}

\item{fun_to_aggregate}{Function that will operate along the nowcast
vectors after summing across delays. Eventually, we can add things like
mean, but for now since we are only providing a negative binomial
observation model, we can only allow sum. Currently supported
functions: \code{sum}.}

\item{k}{Integer indicating the number of reference times to apply the
\code{fun_to_aggregate} over to create target used to compute the nowcast
errors.}
}
\value{
List of matrices labelled as \code{obs} and \code{pred} corresponding to the
observed components that had yet to be observed as of the retrospective
nowcast time.
}
\description{
Extract observations and predictions by retrospective nowcast date and
horizon
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

trunc_rts <- truncate_triangles(triangle, n = 2)
retro_rts <- construct_triangles(trunc_rts)

retro_nowcasts <- fill_triangles(retro_rts, n = 5)
obs_and_pred_list <- extract_obs_and_pred(
  pt_nowcast_matrices = retro_nowcasts,
  trunc_reporting_triangles = trunc_rts,
  retro_reporting_triangles = retro_rts,
  n = 2
)
}
