% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_predictions}
\alias{extract_predictions}
\title{Extract from one matrix only elements that are missing in another}
\usage{
extract_predictions(pt_nowcast_mat, rep_mat)
}
\arguments{
\item{pt_nowcast_mat}{Matrix containing a mix of predicted and observed
values}

\item{rep_mat}{Matrix containing only the observed elements of the
reporting triangle}
}
\value{
\code{pred_mat} Matrix containing the elements from \code{pt_nowcast_mat}
for only the elements that are missing in \code{obs_mat}
}
\description{
Extract from one matrix only elements that are missing in another
}
\examples{
pt_nowcast_mat <- matrix(
  c(
    1, 3, 5, 7,
    4, 7, 8, 9,
    9, 10, 3, 5,
    3, 4, 8, 5
  ),
  nrow = 4,
  byrow = TRUE
)

reporting_matrix <- matrix(
  c(
    1, 3, 5, 7,
    4, 7, 8, NA,
    9, 10, NA, NA,
    3, NA, NA, NA
  ),
  nrow = 4,
  byrow = TRUE
)

other_subset <- extract_predictions(pt_nowcast_mat, reporting_matrix)
other_subset
}
