% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{fit_distribution}
\alias{fit_distribution}
\title{Fit observations vs predictions to an observation model}
\usage{
fit_distribution(obs, pred, observation_model_name = "dnbinom")
}
\arguments{
\item{obs}{Matrix of observations with retrospective nowcast dates as rows
and horizons as columns.}

\item{pred}{Matrix of predictions with retrospective nowcast dates as rows
and horizons as columns.}

\item{observation_model_name}{Character string indicating the choice of
observation model to fit to the predicted nowcasts versus the
observations. Default is \code{"negative binomial"}.}
}
\value{
Vector of parameters corresponding to the chosen
\code{observation_model} of length of the number of columns in \code{obs} and
\code{pred}, with each element corresponding to a horizon.
}
\description{
Fit observations vs predictions to an observation model
}
\examples{
obs <- matrix(c(
  4, 5,
  6, 7
), byrow = 2, nrow = 2)
pred <- matrix(c(
  3.7, 6.1,
  5.2, 10.4
), byrow = 2, nrow = 2)
disp <- fit_distribution(obs, pred,
  observation_model_name = "negative binomial"
)
disp
}
