% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_obs_vs_pred.R
\name{fit_obs_vs_pred}
\alias{fit_obs_vs_pred}
\title{Fit observations vs predictions to an observation model}
\usage{
fit_obs_vs_pred(obs, pred, observation_model = "dnbinom")
}
\arguments{
\item{obs}{Matrix of observations with retrospective nowcast dates as rows
and horizons as columns.}

\item{pred}{Matrix of predictions with retrospective nowcast dates as rows
and horizons as columns.}

\item{observation_model}{Character string indicating the choice of observation model
to fit to the predicted nowcasts versus the observations. Default is
\verb{negative binomial}.}
}
\value{
Vector of parameters corresponding to the chosen
\code{observation_model} of length of the number of columns in \code{obs} and
\code{pred}, with each element corresponding to a horizon.
}
\description{
Fit observations vs predictions to an observation model
}
\examples{
obs <- matrix(
c(4, 15, 10,
 8, 10, 6
 ),
 nrow = 2,
 byrow = TRUE
 )
pred <- matrix(
 c(11, 21, 8,
 10, 15.3, 3.5
 ),
  nrow = 2,
  byrow = TRUE
 )
disp <- fit_obs_vs_pred(obs, pred, observation_model = "negative binomial")
}
