% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nowcasts.R
\name{generate_nowcasts}
\alias{generate_nowcasts}
\title{Generate retrospective nowcasts}
\usage{
generate_nowcasts(list_of_rts, n = min(sapply(list_of_rts, nrow)))
}
\arguments{
\item{list_of_rts}{List of reporting triangle matrices, in order
from most recent (most complete) to least recent. Bottom right of the
matrices should contain NAs.}

\item{n}{Integer indicating the number of observations
(number of rows) to use to estimate the delay distribution for each
reporting triangle. Default is the minimum of the number of rows of
all the matrices in the \code{list_of_rts}}
}
\value{
List of the same number of elements as the input \code{list_of_rts}
but with each reporting triangle filled in based on the delay estimated
in that reporting triangle.
}
\description{
This function ingests a list of retrospective reporting triangles and
generates a list of retrospective reporting squares, or "complete"
reporting triangles. It uses the specified \code{n_history_delay}  number of
observations to estimate the empirical delay for each retrospective
reporting triangle.
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

retro_rts <- generate_triangles(
  triangle = triangle,
  n = 2
)
retro_nowcasts <- generate_nowcasts(
  list_of_rts = retro_rts,
  n = 5
)
}
