% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nowcasts.R
\name{generate_point_nowcast}
\alias{generate_point_nowcast}
\title{Generate point nowcast}
\usage{
generate_point_nowcast(
  triangle_to_nowcast,
  max_delay = ncol(triangle_to_nowcast) - 1,
  n = nrow(triangle_to_nowcast),
  delay_pmf = NULL
)
}
\arguments{
\item{max_delay}{Integer indicating the maximum delay to estimate, in units
of the delay. The default is to use the whole reporting triangle,
\code{ncol(triangle) -1}.}

\item{n}{Integer indicating the number of reference times (observations) to
be used in the estimate of the reporting delay, always starting from the
most recent reporting delay. The default is to use the whole reporting
triangle, so \code{nrow(triangle)}}
}
\value{
\code{comp_rep_square} Matrix of the same number of rows and columns as
the \code{triangle_to_nowcast} but with the missing values filled in as point
estimates
}
\description{
This function ingests a reporting triangle matrix and optionally, a delay
distribution, and returns a completed reporting square which represents
the point nowcast. If a delay distribution is specified, this will be
used to generate the nowcast, otherwise, a delay distribution will be
estimated from the \code{triangle_to_nowcast}.
}
\examples{
triangle <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, NA,
    80, 40, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 5,
  byrow = TRUE
)
reporting_square <- generate_point_nowcast(
  triangle_to_nowcast = triangle
)
print(reporting_sqaure)
}
