% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pt_nowcast_mat_list.R
\name{generate_pt_nowcast_mat}
\alias{generate_pt_nowcast_mat}
\title{Generate point nowcast}
\usage{
generate_pt_nowcast_mat(
  reporting_triangle,
  max_delay = ncol(reporting_triangle) - 1,
  n = nrow(reporting_triangle),
  delay_pmf = NULL
)
}
\arguments{
\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{max_delay}{Integer indicating the maximum delay to estimate, in units
of the delay. The default is to use the whole reporting triangle,
\code{ncol(triangle) -1}.}

\item{n}{Integer indicating the number of reference times (observations) to
be used in the estimate of the reporting delay, always starting from the
most recent reporting delay. The default is to use the whole reporting
triangle, so \code{nrow(triangle)}.}

\item{delay_pmf}{Vector of delays assumed to be indexed starting at the
first delay column in \code{triangle_to_nowcast}.}
}
\value{
\code{point_nowcast_matrix} Matrix of the same number of rows and
columns as the \code{reporting_triangle} but with the missing values filled
in as point estimates.
}
\description{
This function ingests a reporting triangle matrix and optionally, a delay
distribution, and returns a completed reporting square which represents
the point nowcast. If a delay distribution is specified, this will be
used to generate the nowcast, otherwise, a delay distribution will be
estimated from the \code{reporting_triangle}.
}
\examples{
triangle <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, NA,
    80, 40, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 5,
  byrow = TRUE
)
point_nowcast_matrix <- generate_pt_nowcast_mat(
  reporting_triangle = triangle
)
print(point_nowcast_matrix)
}
