% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_retro_triangles.R
\name{generate_retro_triangle}
\alias{generate_retro_triangle}
\title{Get a single retrospective triangle}
\usage{
generate_retro_triangle(t, matr_observed)
}
\arguments{
\item{t}{Integer indicating the number of days prior to generate the
retrospective reporting triangle for}

\item{matr_observed}{Matrix of the incomplete reporting triangle
to be used to generate retrospective nowcasts, with rows representing the
time points of reference and columns representing the delays}
}
\value{
Matrix with \code{t} fewer rows than \code{matr_observed}, replicating what
would have been observed as of \code{t} days prior.
}
\description{
This function takes in a integer \code{t} and a reporting triangle and generates
the reporting triangle that would have been observed as of \code{t} units
earlier (starting from the bottom of the reporting triangle)
}
