% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_retro_triangles.R
\name{generate_retro_triangles}
\alias{generate_retro_triangles}
\title{Generate retrospective reporting triangles}
\usage{
generate_retro_triangles(triangle, n = nrow(triangle) - ncol(triangle) - 1)

}
\arguments{
\item{triangle}{Matrix of the reporting triangle/rectangle
to be used to generate retrospective triangles, with rows representing the
time points of reference and columns representing the delays.}

\item{n}{Integer indicating the number of retrospective
reporting triangles to be generated, always starting from the most
recent reference time. Default is  only generate truncated matrices
that have sufficient number of rows to generate a nowcast from, though
any number can be specified.}
}
\value{
A list of \code{n} retrospective reporting triangle matrices.
with as many rows as available given the truncation, and the same number
of columns as \code{triangle}.
}
\description{
This function ingests a reporting triangle/matrix and the number of
retrospective reporting triangles we want to create \code{n}, and iteratively
generates the reporting triangle that would have been available as of the
maximum reference time, working from bottom to top for \code{n} snapshots
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

retro_rts <- generate_retro_triangles(
  triangle = triangle,
  n = 2
)
print(retro_rts[[1]])
print(retro_rts[[2]])
}
