% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-getters.R
\name{get_delays_from_dates}
\alias{get_delays_from_dates}
\title{Compute delays between report dates and reference dates}
\usage{
get_delays_from_dates(report_dates, reference_dates, delays_unit)
}
\arguments{
\item{report_dates}{Date vector of report dates.}

\item{reference_dates}{Date vector of reference dates.}

\item{delays_unit}{Character string specifying the temporal granularity of
the delays. Options are \code{"days"}, \code{"weeks"}, \code{"months"}, \code{"years"}.
For the matrix method, this is simply passed as an item in the
\code{reporting_triangle} object which will later be used to create a nowcast
data.frame. For the data.frame method, this is used to compute the delay
in terms of the specified unit, and to expand the combinations of
reference dates and delays to the complete set of combinations in
the reporting triangle.  Default is \code{"days"}.}
}
\value{
Numeric vector of delays.
}
\description{
Computes delays between report dates and reference dates
using the specified time unit. This is the inverse operation of
\code{\link[=get_report_dates]{get_report_dates()}}.
}
\examples{
# Compute delays in days
ref_date <- as.Date("2024-01-01")
report_date <- as.Date("2024-01-08")
get_delays_from_dates(report_date, ref_date, "days") # 7

# Compute delays in weeks
report_date_weeks <- as.Date("2024-01-15")
get_delays_from_dates(report_date_weeks, ref_date, "weeks") # 2
}
\seealso{
Reporting triangle construction and validation
\code{\link{[<-.reporting_triangle}()},
\code{\link{as.data.frame.reporting_triangle}()},
\code{\link{as.matrix.reporting_triangle}()},
\code{\link{as_ChainLadder_triangle}()},
\code{\link{as_reporting_triangle}()},
\code{\link{as_reporting_triangle.data.frame}()},
\code{\link{as_reporting_triangle.matrix}()},
\code{\link{as_reporting_triangle.triangle}()},
\code{\link{assert_reporting_triangle}()},
\code{\link{get_delays_unit}()},
\code{\link{get_max_delay}()},
\code{\link{get_mean_delay}()},
\code{\link{get_quantile_delay}()},
\code{\link{get_reference_dates}()},
\code{\link{get_report_dates}()},
\code{\link{get_reporting_structure}()},
\code{\link{head.reporting_triangle}()},
\code{\link{is_reporting_triangle}()},
\code{\link{new_reporting_triangle}()},
\code{\link{print.reporting_triangle}()},
\code{\link{reporting_triangle-class}},
\code{\link{summary.reporting_triangle}()},
\code{\link{tail.reporting_triangle}()},
\code{\link{truncate_to_delay}()},
\code{\link{truncate_to_quantile}()},
\code{\link{validate_reporting_triangle}()}
}
\concept{reporting_triangle}
