% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nowcast_pred_draws.R
\name{get_nowcast_draws}
\alias{get_nowcast_draws}
\title{Generate multiple draws of a nowcast combining observed and predicted values}
\usage{
get_nowcast_draws(
  point_nowcast_matrix,
  reporting_triangle,
  dispersion,
  draws = 1000
)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of point nowcast predictions and
observations, with rows representing the reference times and columns
representing the delays.}

\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{dispersion}{Vector of dispersion parameters indexed by horizon from
minus one to the maximum delay.}

\item{draws}{Integer indicating the number of draws of the predicted
nowcast vector to generate. Default is \code{1000}.}
}
\value{
Dataframe containing information for multiple draws with columns
for the reference time (\code{time}), the predicted counts (\code{pred_count}), and
the draw number (\code{draw}).
}
\description{
Generate multiple draws of a nowcast combining observed and predicted values
}
\examples{
point_nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 16.8,
    80, 40, 21.2, 19.5,
    70, 34.5, 15.4, 9.1
  ),
  nrow = 5,
  byrow = TRUE
)
reporting_triangle <- generate_triangle(point_nowcast_matrix)
disp <- c(0.8, 12.4, 9.1)
nowcast_draws <- get_nowcast_draws(
  point_nowcast_matrix,
  reporting_triangle,
  disp,
  draws = 5
)
nowcast_draws
}
