% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_uncertainty.R
\name{get_nowcast_mat_draw}
\alias{get_nowcast_mat_draw}
\title{Get a probabilistic draw from the observation model}
\usage{
get_nowcast_mat_draw(point_nowcast_matrix, disp)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of both observations and point estimates
of counts at each reference time and delay.}

\item{disp}{Vector of dispersion parameters indexed starting at delay = 1,}
}
\value{
\code{nowcast_w_obs_error} Matrix containing the same upper left values
as the \code{point_nowcast_matrix}, with the bottom right containing
probabilistic draws assuming a negative binomial
observation model with a mean given by the lower right entries in
\code{point_nowcast_matrix} and dispersion parameters
from \code{disp}
}
\description{
The function ingests an estimate of a point nowcast, in the form of a
point nowcast matrix and a vector of dispersion parameters corresponding to
the dispersion parameter for each delay, and draws
from a negative binomial to get a single expected observed probabilistic
draw of a nowcast
}
\examples{
point_nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 17.8,
    80, 40, 23.2, 15.9,
    70, 35, 20.3, 19.9
  ),
  nrow = 5,
  byrow = TRUE
)

exp_obs_nowcast <- get_nowcast_mat_draw(
  point_nowcast_matrix = point_nowcast_matrix,
  disp = c(8, 1.4, 4)
)
print(exp_obs_nowcast)
}
