% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nowcast_pred_draw.R
\name{get_nowcast_pred_draws}
\alias{get_nowcast_pred_draws}
\title{Get a dataframe of multiple draws of only the predicted elements of the
nowcast vector}
\usage{
get_nowcast_pred_draws(point_nowcast_pred_matrix, disp, n_draws = 1000)
}
\arguments{
\item{point_nowcast_pred_matrix}{Matrix containing only the elements that
have not yet been observed as of the final reference date}

\item{disp}{Vector of dispersion parameters indexed by horizon from minus
one to the maximum delay}

\item{n_draws}{Integer indicating the number of draws of the predicted
nowcast vector to generate. Default is \code{1000}.}
}
\value{
Dataframe containing the predicted point nowcast vectors indexed by
reference time (\code{pred_count}), reference time (\code{time}), and the draw index
(\code{draw}).
}
\description{
Get a dataframe of multiple draws of only the predicted elements of the
nowcast vector
}
\examples{
point_nowcast_pred_matrix <-
matrix(
  c(
    NA, NA, NA, NA,
    NA, NA, NA, NA,
    NA, NA, NA, 16.8,
    NA, NA, 21.2, 19.5,
    NA, 34.5, 15.4, 9.1
  ),
  nrow = 4,
  byrow = TRUE
)
disp <- c(0.8, 12.4, 9.1)
nowcast_pred_draws_df <- get_nowcast_pred_draw(
                         point_nowcast_pred_matrix,
                         disp,
                         500)
nowcast_pred_draws_df
}
