% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle_df-getters.R
\name{get_strata}
\alias{get_strata}
\title{Get strata column names from reporting_triangle_df}
\usage{
get_strata(x)
}
\arguments{
\item{x}{A \link{reporting_triangle_df} object}
}
\value{
Character vector of strata column names, or NULL if no strata
}
\description{
Get strata column names from reporting_triangle_df
}
\examples{
# No strata
rt_df <- as_reporting_triangle_df(syn_nssp_df[syn_nssp_df$report_date <= "2026-04-01", ])
get_strata(rt_df) # NULL

# With strata
rt_df_strata <- as_reporting_triangle_df(
  germany_covid19_hosp,
  by = c("age_group", "location")
)
get_strata(rt_df_strata) # c("age_group", "location")
}
\seealso{
Other reporting_triangle_df: 
\code{\link{[.reporting_triangle_df}()},
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.data.frame}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{assert_reporting_triangle_df}()},
\code{\link{head.reporting_triangle_df}()},
\code{\link{is_reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{print.reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{summary.reporting_triangle_df}()},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{truncate_to_delay}()},
\code{\link{truncate_to_quantile}()},
\code{\link{validate_reporting_triangle_df}()}
}
\concept{reporting_triangle_df}
