% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{get_structure}
\alias{get_structure}
\title{Get structure from a reporting triangle}
\usage{
get_structure(x)
}
\arguments{
\item{x}{A \link{reporting_triangle} object.}
}
\value{
Integer or vector specifying the reporting structure.
}
\description{
Get structure from a reporting triangle
}
\seealso{
Reporting triangle construction and validation
\code{\link{as_ChainLadder_triangle}()},
\code{\link{as_reporting_triangle}()},
\code{\link{as_reporting_triangle.data.frame}()},
\code{\link{as_reporting_triangle.matrix}()},
\code{\link{as_reporting_triangle.triangle}()},
\code{\link{assert_reporting_triangle}()},
\code{\link{detect_structure}()},
\code{\link{get_max_delay}()},
\code{\link{new_reporting_triangle}()},
\code{\link{reporting_triangle-class}}
}
\concept{reporting_triangle}
