% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle_df-methods.R
\name{head.reporting_triangle_df}
\alias{head.reporting_triangle_df}
\title{Get first rows of a reporting_triangle_df}
\usage{
\method{head}{reporting_triangle_df}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \link{reporting_triangle_df} object.}

\item{n}{Integer indicating the number of rows to return. Default is 6.}

\item{...}{Additional arguments (not currently used).}
}
\value{
First rows as a reporting_triangle_df.
}
\description{
Get first rows of a reporting_triangle_df
}
\examples{
# Get first 3 rows
rt_df <- as_reporting_triangle_df(syn_nssp_df[syn_nssp_df$report_date <= "2026-04-01", ])
head(rt_df, n = 3)
}
\seealso{
Other reporting_triangle_df:
\code{\link{[.reporting_triangle_df}()},
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.data.frame}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{assert_reporting_triangle_df}()},
\code{\link{get_strata}()},
\code{\link{is_reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{print.reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{summary.reporting_triangle_df}()},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{validate_reporting_triangle_df}()}
}
\concept{reporting_triangle_df}
