% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle_df-class.R
\name{is_reporting_triangle_df}
\alias{is_reporting_triangle_df}
\title{Check if an object is a reporting_triangle_df}
\usage{
is_reporting_triangle_df(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
Logical indicating whether the object is a reporting_triangle_df.
}
\description{
Check if an object is a reporting_triangle_df
}
\seealso{
Other reporting_triangle_df: 
\code{\link{[.reporting_triangle_df}()},
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.data.frame}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{assert_reporting_triangle_df}()},
\code{\link{get_strata}()},
\code{\link{head.reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{print.reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{summary.reporting_triangle_df}()},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{truncate_to_delay}()},
\code{\link{truncate_to_quantile}()},
\code{\link{validate_reporting_triangle_df}()}
}
\concept{reporting_triangle_df}
