% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_matrix_list_to_df.R
\name{nowcast_matrix_to_df}
\alias{nowcast_matrix_to_df}
\title{Convert a reporting square matrix to a long data frame}
\usage{
nowcast_matrix_to_df(matrix, draw = NULL)
}
\arguments{
\item{matrix}{Matrix in the form of a reporting triangle/square, where the
rows indicate the reference time and the columns indicate the delay.}

\item{draw}{integer tracking the draw that the reporting matrix
corresponds to. Default is \code{NULL} which will not return a draw column.}
}
\value{
A long dataframe of the length of the product of the number of
columns and the number of rows, with information on what time and delay
the observation corresponds to.
}
\description{
Convert a reporting square matrix to a long data frame
}
\examples{
nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 18,
    80, 40, 24, 16,
    70, 35, 21, 19,
    67, 34, 15, 9
  ),
  nrow = 6,
  byrow = TRUE
)

long_df <- nowcast_matrix_to_df(nowcast_matrix)
print(long_df)
}
