% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSSP_line_list_raw}
\alias{NSSP_line_list_raw}
\title{A mock dataset resembling line-list (each row is a patient) data from the
National Syndromic Surveillance System (NSSP) accessed via the Essence
platform. All entries are mock data, formatted to look as close to the
real raw data as possible.}
\format{
A data.frame with 25 rows and 8 columns.
\describe{
\item{C_Processed_BioSenseID}{Unique identifier for each patient}
\item{CCDDParsed}{Character string indicating primary symptoms and
corresponding diagnoses codes.}
\item{DischargeDiagnosisMDTUpdates}{Character string formatted as a
dictionary with indices and corresponding time stamps formatted as
YYYY-MM-DD HH:MM:SS.}
\item{DischargeDiagnosisUpdates}{Character string formatted as a
dictionary with indices and corresponding diagnoses codes pertaining to
thie diagnosis associated with that event.}
\item{HasBeenAdmitted}{Numeric indicating whether the patient was
admitted (0 for no, 1 for admission).}
\item{C_Visit_Date_Time}{Date-time indicating the time stamp of the
initial visit in YYYY-MM-DD HH:MM:SS format}
\item{c_race}{Character string indicating the race/ethnicity of the
patient}
\item{sex}{Character string indicating the gender of the patient}
}
}
\source{
Created for package demonstration to provide an example of how to
pre-process this dataset to obtain a reporting triangle.
}
\usage{
NSSP_line_list_raw
}
\description{
A mock dataset resembling line-list (each row is a patient) data from the
National Syndromic Surveillance System (NSSP) accessed via the Essence
platform. All entries are mock data, formatted to look as close to the
real raw data as possible.
}
\keyword{datasets}
