% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_reporting_triangle.R
\name{preprocess_reporting_triangle}
\alias{preprocess_reporting_triangle}
\title{Pre-process the reporting triangle}
\usage{
preprocess_reporting_triangle(triangle, max_delay)
}
\arguments{
\item{triangle}{the reporting triangle as a data.frame, with rows as the reference
date and columns as the delay, in any units. Assumes that the delays will
be in the format \code{{delay}} with no suffix for the unit}

\item{max_delay}{the maximum delay, in the delay units, to filter the
reporting triangle}
}
\value{
trunc_triangle a dataframe in the same format as \code{triangle}, but
truncated to include only the maximum delay number of delay columns
and with negative values of reporting handled via passing them to the
subsequent days delay
}
\description{
Takes in a reporting triangle with the reference date as the first column,
and all subsequent columns named by their delay (unitless, so for example a
delay of 3 days or weeks would have the column name \code{3}), and values
indicating the number of new confirmed cases assigned to that reference
date with that particular delay. It returns a dataframe in the same format
as the input triangle, but truncated to include only the maximum delay
number of columns and with negative values of reporting handled via passing
them to the subsequent days delay.
Modified from https://github.com/KITmetricslab/RESPINOW-Hub/blob/main/code/baseline/functions.R #nolint
}
