% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_opts.R
\name{print.aggregation_opts}
\alias{print.aggregation_opts}
\title{Print Method for Aggregation Options}
\usage{
\method{print}{aggregation_opts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{aggregation_opts}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for Aggregation Options
}
\examples{
agg <- aggregation_observed()
print(agg)

agg_custom <- aggregation_opts(
  ref_time = identity,
  delay = function(x) colSums(x, na.rm = TRUE)
)
print(agg_custom)
}
\seealso{
Other aggregation_opts: 
\code{\link{aggregation_observed}()},
\code{\link{aggregation_opts}()},
\code{\link{assert_aggregation_opts}()},
\code{\link{assert_aggregation_opts.aggregation_opts}()},
\code{\link{assert_aggregation_opts.default}()}
}
\concept{aggregation_opts}
