% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{print.reporting_triangle}
\alias{print.reporting_triangle}
\title{Print a reporting_triangle object}
\usage{
\method{print}{reporting_triangle}(x, ...)
}
\arguments{
\item{x}{A \link{reporting_triangle} object to print.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
Invisibly returns the object.
}
\description{
Print a reporting_triangle object
}
\seealso{
Reporting triangle construction and validation
\code{\link{as_ChainLadder_triangle}()},
\code{\link{as_reporting_triangle}()},
\code{\link{as_reporting_triangle.data.frame}()},
\code{\link{as_reporting_triangle.matrix}()},
\code{\link{as_reporting_triangle.triangle}()},
\code{\link{assert_reporting_triangle}()},
\code{\link{detect_structure}()},
\code{\link{new_reporting_triangle}()},
\code{\link{reporting_triangle-class}},
\code{\link{summary.reporting_triangle}()}
}
\concept{reporting_triangle}
