% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle_df-methods.R
\name{print.reporting_triangle_df}
\alias{print.reporting_triangle_df}
\title{Print a reporting_triangle_df object}
\usage{
\method{print}{reporting_triangle_df}(x, n = 10, ...)
}
\arguments{
\item{x}{A \link{reporting_triangle_df} object to print.}

\item{n}{Maximum number of rows to display. Default is 10.
Set to NULL to display all rows.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
Invisibly returns the object.
}
\description{
Print a reporting_triangle_df object
}
\seealso{
Other reporting_triangle_df: 
\code{\link{[.reporting_triangle_df}()},
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.data.frame}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{assert_reporting_triangle_df}()},
\code{\link{get_strata}()},
\code{\link{head.reporting_triangle_df}()},
\code{\link{is_reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{summary.reporting_triangle_df}()},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{truncate_to_delay}()},
\code{\link{truncate_to_quantile}()},
\code{\link{validate_reporting_triangle_df}()}
}
\concept{reporting_triangle_df}
