% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_model.R
\name{print.uncertainty_model}
\alias{print.uncertainty_model}
\title{Print Method for Uncertainty Models}
\usage{
\method{print}{uncertainty_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{uncertainty_model}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for Uncertainty Models
}
\examples{
model <- uncertainty_nb()
print(model)
}
\seealso{
Other uncertainty_model: 
\code{\link{assert_uncertainty_model}()},
\code{\link{assert_uncertainty_model.default}()},
\code{\link{assert_uncertainty_model.uncertainty_model}()},
\code{\link{assert_uncertainty_model.uncertainty_nb}()},
\code{\link{assert_uncertainty_model.uncertainty_poisson}()},
\code{\link{uncertainty_model}()}
}
\concept{uncertainty_model}
