% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_opts.R
\name{print.uncertainty_opts}
\alias{print.uncertainty_opts}
\title{Print Method for Uncertainty Options}
\usage{
\method{print}{uncertainty_opts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{uncertainty_opts}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for Uncertainty Options
}
\examples{
opts <- uncertainty_opts()
print(opts)

opts_pois <- uncertainty_opts(model = uncertainty_poisson())
print(opts_pois)
}
\seealso{
Other uncertainty_opts: 
\code{\link{uncertainty_opts}()}
}
\concept{uncertainty_opts}
