% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_strategy.R
\name{print.uncertainty_strategy}
\alias{print.uncertainty_strategy}
\title{Print Method for Uncertainty Strategies}
\usage{
\method{print}{uncertainty_strategy}(x, ...)
}
\arguments{
\item{x}{An object of class \code{uncertainty_strategy}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for Uncertainty Strategies
}
\examples{
strategy <- uncertainty_by_horizon()
print(strategy)
}
\seealso{
Other uncertainty_strategy: 
\code{\link{uncertainty_by_horizon}()},
\code{\link{uncertainty_strategy}()}
}
\concept{uncertainty_strategy}
