% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{ref_time_aggregator_default}
\alias{ref_time_aggregator_default}
\title{Reference time aggregator default function}
\usage{
ref_time_aggregator_default(fun = identity, ...)
}
\arguments{
\item{fun}{Function to aggregate across reference times of the reporting
triangle, default is \code{identity} which returns what is passed in.}

\item{...}{Additional arguments passed to the function}
}
\value{
Object of class aggregator containing functions and arguments
}
\description{
Reference time aggregator default function
}
\examples{
ref_time_aggregator()
}
