% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_lower_right_with_NA}
\alias{replace_lower_right_with_NA}
\title{Replace the lower right triangle of the matrix with NAs}
\usage{
replace_lower_right_with_NA(matrix)
}
\arguments{
\item{matrix}{Matrix}
}
\value{
A matrix of the same dimensions, with NAs for all the lower right
entries.
}
\description{
Replace the lower right triangle of the matrix with NAs
}
\examples{
triangle_w_zeros <- matrix(
c( 1, 3, 5, 7,
   4, 7, 8, 0,
   9, 10, 0, 0,
   3, 0, 0, 0),
 nrow = 4,
 byrow = TRUE)

rep_tri <- replace_lower_right_with_NA(triangle_w_zeros)
print(rep_tri)

}
