% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-methods.R
\name{[.reporting_triangle}
\alias{[.reporting_triangle}
\title{Subset reporting_triangle objects}
\usage{
\method{[}{reporting_triangle}(x, ...)
}
\arguments{
\item{x}{A \link{reporting_triangle} object}

\item{...}{Indices for subsetting}
}
\value{
A \link{reporting_triangle} object with the subset data
}
\description{
Extract or replace parts of a reporting_triangle object while preserving its
attributes.
}
