% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle_df-methods.R
\name{[.reporting_triangle_df}
\alias{[.reporting_triangle_df}
\title{Subset reporting_triangle_df objects}
\usage{
\method{[}{reporting_triangle_df}(x, ...)
}
\arguments{
\item{x}{A \link{reporting_triangle_df} object}

\item{...}{Indices for subsetting}
}
\value{
A \link{reporting_triangle_df} object with the subset data
}
\description{
Extract parts of a reporting_triangle_df object while preserving its
attributes.
}
\seealso{
Other reporting_triangle_df: 
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.data.frame}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{assert_reporting_triangle_df}()},
\code{\link{get_strata}()},
\code{\link{head.reporting_triangle_df}()},
\code{\link{is_reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{print.reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{summary.reporting_triangle_df}()},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{validate_reporting_triangle_df}()}
}
\concept{reporting_triangle_df}
