% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{[<-.reporting_triangle}
\alias{[<-.reporting_triangle}
\title{Subset assignment for reporting_triangle objects}
\usage{
\method{[}{reporting_triangle}(x, ...) <- value
}
\arguments{
\item{x}{A \link{reporting_triangle} object.}

\item{...}{Row and column indices.}

\item{value}{Values to assign.}
}
\value{
The modified reporting_triangle object.
}
\description{
Assignment method that validates the result after modification.
}
\keyword{internal}
