% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle_df-methods.R
\name{summary.reporting_triangle_df}
\alias{summary.reporting_triangle_df}
\title{Summarize a reporting_triangle_df object}
\usage{
\method{summary}{reporting_triangle_df}(object, ...)
}
\arguments{
\item{object}{A \link{reporting_triangle_df} object to summarize.}

\item{...}{Additional arguments not used.}
}
\value{
Invisibly returns the object.
}
\description{
Summarize a reporting_triangle_df object
}
\examples{
# Display summary statistics
rt_df <- as_reporting_triangle_df(syn_nssp_df[syn_nssp_df$report_date <= "2026-04-01", ])
summary(rt_df)
}
\seealso{
Other reporting_triangle_df:
\code{\link{[.reporting_triangle_df}()},
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.data.frame}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{assert_reporting_triangle_df}()},
\code{\link{get_strata}()},
\code{\link{head.reporting_triangle_df}()},
\code{\link{is_reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{print.reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{validate_reporting_triangle_df}()}
}
\concept{reporting_triangle_df}
