% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{syn_nssp_rep_tri_df}
\alias{syn_nssp_rep_tri_df}
\title{A synthetic dataset containing the number of incident cases indexed by
reference date and report date.}
\format{
A data.frame with 3795 rows and 3 columns.
\describe{
\item{reference_date}{Date of the primary event occurred (e.g. date of
hospital admissions, specimen collection date, symptom onset), formatted
in ISO8601 standards as YYYY-MM-DD.}
\item{report_date}{Date the event was reported into the surveillance
system, formatted as YYYY-MM-DD.}
\item{count}{Number of incident events (e.g. cases) occurring on the
specified reference date and reported on the report date.}
}
}
\source{
Created for package demonstration.
}
\usage{
syn_nssp_rep_tri_df
}
\description{
A synthetic dataset containing the number of incident cases indexed by
reference date and report date.
}
\keyword{datasets}
