% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_to_rows.R
\name{truncate_to_rows}
\alias{truncate_to_rows}
\title{Truncate reporting triangle by removing bottom rows}
\usage{
truncate_to_rows(
  reporting_triangle,
  n = nrow(reporting_triangle) - sum(is.na(rowSums(reporting_triangle))) - 1
)
}
\arguments{
\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{n}{Integer indicating the number of retrospective
truncated triangles to be generated, always starting from the most
recent reference time. Default is to generate truncated matrices for
each row up until there are insufficient rows to generate nowcasts
from, where the minimum requirement is one more than the  number of
horizon rows (rows containing NAs).}
}
\value{
\code{trunc_rep_tri_list} List of \code{n} truncated reporting triangle
matrices with as many rows as available given the truncation, and the same
number of columns as \code{reporting_triangle}.
}
\description{
Generates a list of retrospective reporting triangles by successively
removing rows from the bottom of the original triangle.
Each truncated triangle represents what would have been observed at an
earlier reference time.
This function truncates by row count, removing the most recent observations.
For other truncation approaches, see \code{\link[=truncate_to_quantile]{truncate_to_quantile()}}.
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

truncated_rts <- truncate_to_rows(triangle, n = 2)
truncated_rts[1:2]
}
\seealso{
Retrospective data generation functions
\code{\link{construct_triangle}()},
\code{\link{construct_triangles}()},
\code{\link{truncate_to_row}()}
}
\concept{generate_retrospective_data}
