% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle_df-class.R
\name{validate_reporting_triangle_df}
\alias{validate_reporting_triangle_df}
\title{Validate a reporting_triangle_df object}
\usage{
validate_reporting_triangle_df(data)
}
\arguments{
\item{data}{A \link{reporting_triangle_df} object to validate}
}
\value{
The validated object (invisibly) or throws error
}
\description{
Validate a reporting_triangle_df object
}
\seealso{
Other reporting_triangle_df: 
\code{\link{[.reporting_triangle_df}()},
\code{\link{as.data.frame.reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df}()},
\code{\link{as_reporting_triangle_df.data.frame}()},
\code{\link{as_reporting_triangle_df.reporting_triangle}()},
\code{\link{as_reporting_triangles}()},
\code{\link{assert_reporting_triangle_df}()},
\code{\link{get_strata}()},
\code{\link{head.reporting_triangle_df}()},
\code{\link{is_reporting_triangle_df}()},
\code{\link{new_reporting_triangle_df}()},
\code{\link{print.reporting_triangle_df}()},
\code{\link{reporting_triangle_df-class}},
\code{\link{summary.reporting_triangle_df}()},
\code{\link{tail.reporting_triangle_df}()},
\code{\link{truncate_to_delay}()},
\code{\link{truncate_to_quantile}()}
}
\concept{reporting_triangle_df}
