% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_triangle}
\alias{validate_triangle}
\title{Validate triangle}
\usage{
validate_triangle(triangle, max_delay, n_history)
}
\arguments{
\item{triangle}{a matrix of values with rows indicating the time points and
columns indicating the delays}

\item{max_delay}{integer indicating the maximum delay to estimate, in units
of the delay. The default is to use the whole reporting triangle,
\code{ncol(triangle) -1}.}

\item{n_history}{integer indicating the number of reference dates to be
used in the estimate of the reporting delay, always starting from the most
recent reporting delay. The default is to use the whole reporting triangle,
so \code{nrow(triangle)-1}}
}
\value{
NULL, invisibly
}
\description{
Various checks to make sure that the reporting triangle passed in to
\code{estimate_delay()} is formatted properly.
}
