% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{as_reporting_triangle.matrix}
\alias{as_reporting_triangle.matrix}
\title{Create a \code{reporting_triangle} from a matrix}
\usage{
\method{as_reporting_triangle}{matrix}(
  data,
  max_delay,
  strata = NULL,
  delays_unit = "days",
  reference_dates,
  ...
)
}
\arguments{
\item{data}{Matrix of a reporting triangle where rows are reference times,
columns are delays, and entries are the incident counts.}

\item{max_delay}{Integer indicating the maximum delay.}

\item{strata}{Character string indicating the metadata on the strata of this
reporting triangle. Default is \code{NULL}.}

\item{delays_unit}{Character string specifying the time units to use.
Default is \code{"days"}.}

\item{reference_dates}{Vector of character strings indicating the reference
dates corresponding to each row of the reporting triangle matrix (\code{data}).}

\item{...}{Additional arguments passed to methods.}
}
\description{
This method takes a matrix in the format of a reporting triangle, with rows
as reference dates and columns as delays and elements as incident case
counts and creates a \code{reporting_triangle} object. See other
methods for other data input options.
}
\seealso{
\code{\link{as_reporting_triangle.data.frame}}
\code{\link{as_reporting_triangle}}
}
