% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{as_reporting_triangle}
\alias{as_reporting_triangle}
\title{Create a \code{reporting_triangle} object}
\usage{
as_reporting_triangle(
  data,
  max_delay,
  strata = NULL,
  delays_unit = "days",
  ...
)
}
\arguments{
\item{data}{Data to be nowcasted.}

\item{max_delay}{Integer indicating the maximum delay.}

\item{strata}{Strata associated with the data to be nowcasted. Default is
\code{NULL}}

\item{delays_unit}{Character string specifying the time units to use.
Default is \code{"days"}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{\link{reporting_triangle}} object
}
\description{
Create a \code{reporting_triangle} object
}
\seealso{
Other reporting_triangle:
\code{\link{as_reporting_triangle.data.frame}()},
\code{\link{as_reporting_triangle.matrix}()},
\code{\link{reporting_triangle-class}}
}
\concept{reporting_triangle}
