% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{as_reporting_triangle}
\alias{as_reporting_triangle}
\title{Create a \code{reporting_triangle} object}
\usage{
as_reporting_triangle(
  data,
  max_delay,
  strata = NULL,
  delays_unit = "days",
  ...
)
}
\arguments{
\item{data}{Data to be nowcasted.}

\item{max_delay}{Integer indicating the maximum delay.}

\item{strata}{Character string indicating the metadata on the strata of this
reporting triangle. Default is \code{NULL}.}

\item{delays_unit}{Character string specifying the time units to use.
Default is \code{"days"}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
\code{reporting_triangle} class object which is a list containing:
\itemize{
\item A matrix with which rows are reference times and columns are delays and
entries are incident cases at each reference time and delay.
\item An integer indicating the maximum delay used to create the reporting
triangle
\item A vector of the same length as the rows of the matrix indicating the
dates corresponding to the reference times in the rows of the reporting
triangle.
\item A character string indicating the strata.
\item A vector indicating the "structure" of the reporting triangle.
\item A character string indicating the unit of the delays.
}

See the corresponding \verb{as_reporting_triangle.<data.type>} functions for
more details on the required input formats.
}
\description{
Create a \code{reporting_triangle} object
}
\seealso{
\code{\link{as_reporting_triangle.data.frame}}
\code{\link{as_reporting_triangle.matrix}}
}
