% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{as_reporting_triangle}
\alias{as_reporting_triangle}
\alias{as_reporting_triangle.data.frame}
\title{Create a reporting triangle from a dataframe}
\usage{
as_reporting_triangle(data, ...)

\method{as_reporting_triangle}{data.frame}(
  data,
  max_delay,
  strata = NULL,
  reference_date = "reference_date",
  report_date = "report_date",
  count = "count",
  delays_unit = "days"
)
}
\arguments{
\item{data}{Data.frame to be converted to a reporting triangle matrix.
Can either be in the long tidy format of counts by reference date
and report date or line list data with individual observations
indexed by their reference date and report date.}

\item{strata}{Character string indicating the metadata on the strata of this
reporting triangle. Default is \code{NULL}.}

\item{delays_unit}{Character string specifying the time units to use.
Default is "daily".}

\item{inheritParams}{estimate_delay}
}
\value{
\code{reporting_triangle} class object which is a list containing:
\itemize{
\item A matrix with which rows are reference times and columns are delays and
entries are incident cases at each reference time and delay.
\item An integer indicating the maximum delay used to create the reporting
triangle
\item A vector of the same length as the rows of the matrix indicating the
dates corresponding to the reference times in the rows of the reporting
triangle.
\item A character string indicating the strata.
\item A vector indicating the "structure" of the reporting triangle.
}
}
\description{
Create a reporting triangle from a dataframe
}
\details{
The input needs to be a data.frame or similar with the following columns:
\itemize{
\item \code{reference_date}: Column of type \code{date} or character with the dates of
the reference event.
\item \code{report_date}: Column of type \code{date} or character with the dates of
report of the reference event.
\item \code{count}: Column of numeric or integer indicating the new confirmed counts
pertaining to that reference and report date.
Additional columns can be included but will not be used. The input
dataframe for this function must contain only a single strata, there can
be no repeated reference dates and report dates.
}
}
\examples{
data_as_of_df <- syn_nssp_df |>
  filter(report_date <= "2026-04-01")
as_reporting_triangle(
  data = data_as_of_df,
  max_delay = 25
)
}
