% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{as_reporting_triangle}
\alias{as_reporting_triangle}
\alias{as_reporting_triangle.data.frame}
\title{Create a reporting triangle from a dataframe}
\usage{
as_reporting_triangle(data, ...)

\method{as_reporting_triangle}{data.frame}(
  data,
  reference_date = NULL,
  report_date = NULL,
  max_delay = NULL,
  ...
)
}
\arguments{
\item{...}{Additional arguments}

\item{reference_date}{Name of the column in \code{data} that contains the
reference date.}

\item{report_date}{Name of the column in \code{data} that contains the report
date.}

\item{max_delay}{Integer indicating the maximum delay of the process which
will dictate the number of columns in the resulting reporting triangle.}
}
\value{
Matrix containing reference times as rows and delays as columns
with entries for the number of new counts at each reference time and
delay.
}
\description{
Create a reporting triangle from a dataframe
}
\details{
The input needs to be a data.frame or similar with the following columns:
\itemize{
\item \code{reference_date}: Column of type \code{date} or character with the dates of
the reference event.
\item \code{report_date}: Column of type \code{date} or character with the dates of
report of the reference event.
\item \code{count}: Column of numeric or integer indicating the new confirmed counts
pertaining to that reference and report date.
}
}
\examples{
long_data <- data.frame(
reference_date = c("2025-01-01", "2025-01-01", "2025-01-01",
  "2025-01-02", "2025-01-02", "2025-01-03"),
report_date = c("2025-01-01", "2025-01-02", "2025-01-03",
   "2025-01-02", "2025-01-03", "2025-01-03"),
count = c(10, 2, 1, 12, 3, 13)
  )
as_reporting_triangle(long_data)
}
