% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast_df-class.R
\name{assert_baselinenowcast_df}
\alias{assert_baselinenowcast_df}
\title{Assert validity of \code{baselinenowcast_df} objects}
\usage{
assert_baselinenowcast_df(data)
}
\arguments{
\item{data}{A \code{\link{baselinenowcast_df}} object to check for validity.}
}
\description{
Assert validity of \code{baselinenowcast_df} objects
}
