% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{assert_reporting_triangle}
\alias{assert_reporting_triangle}
\title{Assert validity of \code{reporting_triangle} objects}
\usage{
assert_reporting_triangle(data)
}
\arguments{
\item{data}{A \link{reporting_triangle} object to check for validity.}
}
\description{
Assert validity of \code{reporting_triangle} objects
}
\seealso{
Reporting triangle construction and validation
\code{\link{as_ChainLadder_triangle}()},
\code{\link{as_reporting_triangle}()},
\code{\link{as_reporting_triangle.data.frame}()},
\code{\link{as_reporting_triangle.matrix}()},
\code{\link{as_reporting_triangle.triangle}()},
\code{\link{get_delay_unit}()},
\code{\link{get_max_delay}()},
\code{\link{get_reporting_structure}()},
\code{\link{get_structure}()},
\code{\link{new_reporting_triangle}()},
\code{\link{reporting_triangle-class}}
}
\concept{reporting_triangle}
