% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{assert_reporting_triangle}
\alias{assert_reporting_triangle}
\title{Assert validity of \code{reporting_triangle} objects}
\usage{
assert_reporting_triangle(data)
}
\arguments{
\item{data}{A \code{\link{reporting_triangle}} object to check for validity.}
}
\description{
Assert validity of \code{reporting_triangle} objects
}
