% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{baselinenowcast}
\alias{baselinenowcast}
\title{Generate a nowcast}
\usage{
baselinenowcast(
  data,
  scale_factor = 3,
  prop_delay = 0.5,
  output_type = c("samples", "point"),
  draws = 1000,
  uncertainty_model = NULL,
  uncertainty_sampler = NULL,
  ...
)
}
\arguments{
\item{data}{Data to be nowcasted}

\item{scale_factor}{Numeric value indicating the multiplicative factor on
the maximum delay to be used for estimation of delay and uncertainty.
Default is \code{3}.}

\item{prop_delay}{Numeric value <1 indicating what proportion of all
reference times in the reporting triangle to be used for delay
estimation. Default is \code{0.5}.}

\item{output_type}{Character string indicating whether the output should be
samples (\code{"samples"})from the estimate with full uncertainty or whether to
return the point estimate (\code{"point"}). Default is \code{"samples"}.}

\item{draws}{Integer indicating the number of probabilistic draws to include
if \code{output_type} is \code{"samples"}. Default is 1000.}

\item{uncertainty_model}{(Deprecated) Function that ingests a matrix of
observations and a matrix of predictions and returns a vector that can
be used to apply uncertainty using the same error model. Use
\code{uncertainty} parameter instead.}

\item{uncertainty_sampler}{(Deprecated) Function that ingests a vector or
matrix of predictions and a vector of uncertainty parameters and
generates draws from the observation model. Use \code{uncertainty} parameter
instead.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Data.frame of class \link{baselinenowcast_df}
}
\description{
This function ingests data to be nowcasted and generates a
a \link{baselinenowcast_df} which contains a probabilistic or point
estimate of the final case counts at each reference date in the \code{data}.
See \code{\link[=baselinenowcast.reporting_triangle]{baselinenowcast.reporting_triangle()}} for details on the
input requirements.
}
\seealso{
Main nowcasting interface functions
\code{\link{assert_baselinenowcast_df}()},
\code{\link{baselinenowcast.data.frame}()},
\code{\link{baselinenowcast.reporting_triangle}()},
\code{\link{baselinenowcast_df-class}},
\code{\link{new_baselinenowcast_df}()}
}
\concept{baselinenowcast_df}
