% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{baselinenowcast.reporting_triangle}
\alias{baselinenowcast.reporting_triangle}
\title{Creating a dataframe of nowcast results from a single reporting
triangle}
\usage{
\method{baselinenowcast}{reporting_triangle}(
  data,
  scale_factor = 3,
  prop_delay = 0.5,
  output_type = "samples",
  draws = 1000,
  uncertainty_model = fit_by_horizon,
  uncertainty_sampler = sample_nb,
  delay_pmf = estimate_delay(data$reporting_triangle_matrix),
  uncertainty_params = NULL,
  ...
)
}
\arguments{
\item{data}{\code{reporting_triangle} class object to be nowcasted.}

\item{output_type}{Character string indicating whether the output should be
samples (\code{"samples"})from the estimate with full uncertainty or whether to
return the point estimate (\code{"point"}). Default is \code{"samples"}.}

\item{draws}{Integer indicating the number of probabilistic draws to include
if \code{output_type} is \code{"samples"}. Default is 1000.}

\item{uncertainty_model}{Function that ingests a matrix of observations and a
matrix of predictions and returns a vector that can be used to
apply uncertainty using the same error model. Default is
\code{fit_by_horizon} with arguments of \code{obs} matrix of observations and
\code{pred} the matrix of predictions that fits each column (horizon)
to a negative binomial observation model by default. The user can
specify a different fitting model by replacing the
\code{fit_model} argument in \code{fit_by_horizon}.}

\item{uncertainty_sampler}{Function that ingests a vector or matrix of
predictions and a vector of uncertainty parameters and generates draws
from the observation model. Default is \code{sample_nb} which expects
arguments \code{pred} for the vector of predictions and uncertainty parameters
for the corresponding vector of uncertainty parameters, and draws from a
negative binomial for each element of the vector.}

\item{delay_pmf}{Vector of delays assumed to be indexed starting at the
first delay column in \code{data$reporting_triangle_matrix}. Default is to
estimate from the reporting triangle matrix in \code{data},
\code{estimate_delay(data$reporting_triangle_matrix)}.}

\item{uncertainty_params}{Vector of uncertainty parameters ordered from
horizon 1 to the maximum horizon. Default is \code{NULL}, which will
result in computing the uncertainty parameters from the \code{data}.}

\item{...}{Additional arguments passed to
\code{\link{estimate_uncertainty}}
and \code{\link{sample_nowcast}}.}
}
\value{
A data.frame of class \code{nowcast_df}
}
\description{
Creating a dataframe of nowcast results from a single reporting
triangle
}
\examples{
data_as_of_df <- syn_nssp_df[syn_nssp_df$report_date <= "2026-04-01", ]
rep_tri <- as_reporting_triangle(
  data = data_as_of_df,
  max_delay = 25
)
nowcast_df <- baselinenowcast(rep_tri)
nowcast_df
}
\seealso{
Other nowcast_df:
\code{\link{baselinenowcast}()}
}
\concept{nowcast_df}
