% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{combine_triangle_dfs}
\alias{combine_triangle_dfs}
\title{Combine triangle data.frames}
\usage{
combine_triangle_dfs(
  data,
  reference_date = "reference_date",
  report_date = "report_date",
  count = "count"
)
}
\arguments{
\item{data}{Data.frame in a long tidy format with counts by reference date
and report date for one or more strata. Must contain the following
columns:
.    - Column of type \code{date} or character with the dates of
the primary event occurrence (reference date).
\itemize{
\item Column of type \code{date} or character with the dates of
report of the primary event (report_date).
\item Column of numeric or integer indicating the new confirmed counts
pertaining to that reference and report date (count).
Additional columns indicating the columns which set the unit of a single
nowcast must be included. The user must specify these columns with the
\code{nowcast_unit} argument, or all columns will be assumed to specify the
unit of a single nowcast.
}}

\item{reference_date}{Character string indicating the name of the
column which represents the reference date, or the date of the primary
event occurrence.}

\item{report_date}{Character string indicating the name of the
column which represents the date the primary event was reported.}

\item{count}{Character string indicating the name of the column
containing the number of incident cases on each reference and report date.}
}
\value{
\code{result} Data.frame with the same column names for reference date,
report date, and case count as in \code{data} but summed across all strata in
the original data.
}
\description{
This function ingests a dataframe with case counts indexed by
reference dates and report dates for multiple strata and sums all the case
counts across the shared set of reference and report dates. It requires that
the strata have some overlapping set of reference and report dates,
otherwise it will return an error to indicate to the user that strata
sharing with the set of strata passed in is not possible. If overlapping
sets of reference and report dates are input, it returns a data.frame with
a single set of counts for all the reference and report date combinations
present in all strata in the original data. Note that this may be a subset
of the unique set of reference and report date combinations in the original
data.
}
\examples{
example_data <- data.frame(
  ref_date = as.Date(c("2021-04-06", "2021-04-06", "2021-04-06", "2021-04-06")), # nolint
  rep_date = as.Date(c("2021-04-08", "2021-04-08", "2021-04-10", "2021-04-10")), # nolint
  location = c("DE", "FR", "DE", "FR"),
  age_group = c("00+", "00+", "00+", "00+"),
  cases = c(50, 30, 20, 40)
)
example_data
combined <- combine_triangle_dfs(
  data = example_data,
  reference_date = "ref_date",
  report_date = "rep_date",
  count = "cases"
)
combined
}
