% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_training_volume.R
\name{.allocate_training_volume}
\alias{.allocate_training_volume}
\title{Allocate training volume based on combination of defaults and user-specified
values for training volume for delay and uncertainty estimation.}
\usage{
.allocate_training_volume(
  n_ref_times,
  max_delay,
  n_history_delay = NULL,
  n_retrospective_nowcasts = NULL
)
}
\arguments{
\item{n_ref_times}{Integer indicating the number of reference times (rows)
in the reporting triangle.}

\item{max_delay}{Integer indicating the maximum delay to estimate, in units
of the delay. The default is to use the whole reporting triangle,
\code{ncol(reporting_triangle) -1}.}

\item{n_history_delay}{Integer indicating the number of reference times
(observations) to be used in the estimate of the reporting delay, always
starting from the most recent reporting delay. Default is NULL, which will
be set internally to 1.5* max_delay.}

\item{n_retrospective_nowcasts}{}
}
\value{
list of n_history_delay and n_retrospective_nowcasts
}
\description{
Given the number of references times, the maximum delay, and
optionally the user-specified number of reference times used for delay
(\code{n_history_delay}) and/ot the number of reference times used as
retrospective nowcasts for uncertainty estimation
(\code{n_retrospective_nowcasts}), allocated the training volume accordingly
or return the allocations specified if feasible.
}
