% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_delay.R
\name{.calc_expectation}
\alias{.calc_expectation}
\title{Calculate the updated rows of the expected nowcasted triangle}
\usage{
.calc_expectation(index, expectation, delay_pmf, delay_cdf, n_rows)
}
\arguments{
\item{index}{Integer indicating the delay index}

\item{expectation}{Matrix of the incomplete reporting matrix}

\item{delay_pmf}{Vector of delays assumed to be indexed starting at the
first delay column in \code{rep_mat_to_nowcast}}

\item{delay_cdf}{Vector of cumulative probabilities of the delay PMF}

\item{n_rows}{Number of rows in the expectation matrix}
}
\value{
Matrix with another set of entries corresponding to the updated
values for the specified rows and column
}
\description{
Calculate the updated rows of the expected nowcasted triangle
}
\keyword{internal}
