% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_na_bottom_right}
\alias{.check_na_bottom_right}
\title{Check if matrix only contains NAs in the bottom right}
\usage{
.check_na_bottom_right(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Boolean indicating whether the matrix only contains NAs in the
bottom right (TRUE if only in botom right, FALSE if elsewhere)
}
\description{
Check if matrix only contains NAs in the bottom right
}
\keyword{internal}
