% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_triangle_dfs.R
\name{.combine_triangle_dfs}
\alias{.combine_triangle_dfs}
\title{Combine triangle data.frames}
\usage{
.combine_triangle_dfs(data)
}
\arguments{
\item{data}{Data.frame containing the incident count of cases by reference
date and report date for one or more strata.}
}
\value{
\code{result} Data.frame with the same column names for reference date,
report date, and case count as in \code{data} but summed across all strata in
the original data.
}
\description{
This function ingests a dataframe with case counts indexed by
reference dates and report dates for one or more strata and sums all the case
counts across the shared set of reference and report dates. Note that this
may be a subset of the unique set of reference and report date combinations
in the original data.
}
