% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{.combine_triangle_dfs}
\alias{.combine_triangle_dfs}
\title{Combine triangle data.frames}
\usage{
.combine_triangle_dfs(data, strata_cols = NULL)
}
\arguments{
\item{data}{Data.frame in a long tidy format with counts by reference date
and report date for one or more strata. Must contain the following
columns:
.    - Column of type \code{date} or character with the dates of
the primary event occurrence (reference date).
\itemize{
\item Column of type \code{date} or character with the dates of
report of the primary event (report_date).
\item Column of numeric or integer indicating the new confirmed counts
pertaining to that reference and report date (count).
Additional columns indicating the columns which set the unit of a single
can be included. The user can specify these columns with the
\code{strata_cols} argument, otherwise it will be assumed that the \code{data}
contains only data for a single strata.
}}

\item{strata_cols}{Vector of character strings indicating the names of the
columns in \code{data} that determine how to stratify the data for nowcasting.
The unique combinations of the entries in the \code{strata_cols} denote the
unit of a single nowcast. Within a strata, there can be no repeated
unique combinations of reference dates and report dates. Default is \code{NULL}
which assumes that the data.frame being passed in represents a single
strata (only one nowcast will be produced). All columns that are not
part of the \code{strata_cols} will be removed.}
}
\value{
\code{result} Data.frame with the same column names for reference date,
report date, and case count as in \code{data} but summed across all strata in
the original data.
}
\description{
This function ingests a dataframe with case counts indexed by
reference dates and report dates for multiple strata and sums all the case
counts across the shared set of reference and report dates. It requires that
the strata have some overlapping set of reference and report dates,
otherwise it will return an error to indicate to the user that strata
sharing with the set of strata passed in is not possible. If overlapping
sets of reference and report dates are input, it returns a data.frame with
a single set of counts for all the reference and report date combinations
present in all strata in the original data. Note that this may be a subset
of the unique set of reference and report date combinations in the original
data.
}
