% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_predictions.R
\name{.extract_predictions}
\alias{.extract_predictions}
\title{Extract from one matrix only elements that are missing in another}
\usage{
.extract_predictions(point_nowcast_matrix, reporting_triangle)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of point nowcast predictions and
observations, with rows representing the reference times and columns
representing the delays.}

\item{reporting_triangle}{A \link{reporting_triangle} object with rows
representing reference times and columns representing delays.
Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row (e.g., weekly reporting of daily data).}
}
\value{
Matrix containing the elements from \code{point_nowcast_matrix} for
only the elements that are missing in \code{reporting_triangle}
}
\description{
Extract from one matrix only elements that are missing in another
}
\keyword{internal}
