% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_retrospective_data.R
\name{.get_retro_matrix}
\alias{.get_retro_matrix}
\title{Get retrospective matrix}
\usage{
.get_retro_matrix(t, matr_observed, n_history_delay)
}
\arguments{
\item{t}{Integer indicating the number of days prior to generate the
retrospective reporting triangle for}

\item{matr_observed}{Matrix of the incomplete reporting triangle
to be used to generate retrospective nowcasts, with rows representing the
time points of reference and columns representing the delays}

\item{n_history_delay}{Integer indicating the number of observations in
in each retrospective reporting triangle}
}
\value{
Matrix with \code{t} fewer rows than \code{matr_observed}, replicating what
would have been observed as of \code{t} days prior.
}
\description{
This function takes in a integer t and a reporting triangle and generates
the reporting triangle that would have been observed as of \code{t} days
earlier
}
\keyword{internal}
