% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_reference_times.R
\name{.handle_output_msgs}
\alias{.handle_output_msgs}
\title{Handle output messages after allocation has occurred}
\usage{
.handle_output_msgs(
  n_history_delay,
  n_retrospective_nowcasts,
  size_used,
  prop_delay
)
}
\arguments{
\item{n_history_delay}{Integer indicating the number of reference times
(observations) to be used in the estimate of the reporting delay, always
starting from the most recent reporting delay. Default is NULL, which will
be set internally based on the defaults in \code{allocate_reference_times()}}

\item{n_retrospective_nowcasts}{Integer indicating the number of
retrospective nowcast times to use for uncertainty estimation. Default is
NULL, which will be set internally based on the defaults in
\code{allocate_reference_times()}}

\item{size_used}{Integer indicating number reference times that will be used.}

\item{prop_delay}{Numeric value <1 indicating what proportion of all
reference times in the reporting triangle to be used for delay
estimation. Default is \code{0.5}.}
}
\value{
NULL, invisibly
}
\description{
Handle output messages after allocation has occurred
}
