% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.handle_target_insufficient}
\alias{.handle_target_insufficient}
\title{Helper for when the target is less than the required minimum}
\usage{
.handle_target_insufficient(
  n_target,
  n_required,
  n_min_delay,
  n_min_retro_nowcasts,
  scale_factor,
  max_delay
)
}
\arguments{
\item{n_target}{Integer indicating the target number of reference times.}

\item{n_required}{Integer indicating the number need for both delay and
uncertainty}

\item{n_min_delay}{Integer indicating the number needed for delay
estimation.}

\item{n_min_retro_nowcasts}{Integer indicating the number needed for
uncertainty estimation.}

\item{scale_factor}{Numeric value indicating the multiplicative factor on
the maximum delay to be used for estimation of delay and uncertainty.
Default is \code{3}.}
}
\value{
NULL invisibly
}
\description{
Helper for when the target is less than the required minimum
}
\keyword{internal}
