% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reviser_vintages.R
\name{.infer_delays_unit}
\alias{.infer_delays_unit}
\title{Infer \code{delays_unit} from the spacing of \code{time} values}
\usage{
.infer_delays_unit(times)
}
\arguments{
\item{times}{Vector of dates (or coercible to dates).}
}
\value{
A character string, one of \code{"days"} or \code{"weeks"}.
}
\description{
Infer \code{delays_unit} from the spacing of \code{time} values
}
\keyword{internal}
