% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safelydoesit}
\alias{.safelydoesit}
\title{Safe iterator}
\usage{
.safelydoesit(fun)
}
\arguments{
\item{fun}{Function to wrap around}
}
\value{
Function that will return a NULL if an error occurs
}
\description{
Safe iterator
}
