% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_rep_tri_args}
\alias{.validate_rep_tri_args}
\title{Validate constructor inputs for reporting triangle}
\usage{
.validate_rep_tri_args(reporting_triangle_matrix, reference_dates, delays_unit)
}
\arguments{
\item{reporting_triangle_matrix}{Matrix of reporting triangle}

\item{reference_dates}{Vector of character strings indicating the reference
dates corresponding to each row of the reporting triangle matrix (\code{data}).}

\item{delays_unit}{Character string specifying the temporal granularity of
the delays. Options are \code{"days"}, \code{"weeks"}, \code{"months"}, \code{"years"}.
For the matrix method, this is simply passed as an item in the
\code{reporting_triangle} object which will later be used to create a nowcast
data.frame. For the data.frame method, this is used to compute the delay
in terms of the specified unit, and to expand the combinations of
reference dates and delays to the complete set of combinations in
the reporting triangle.  Default is \code{"days"}.}
}
\value{
NULL invisibly
}
\description{
Validate constructor inputs for reporting triangle
}
\keyword{internal}
