% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{.validate_rep_tri_args}
\alias{.validate_rep_tri_args}
\title{Validate reporting_triangle constructor arguments}
\usage{
.validate_rep_tri_args(reporting_triangle_matrix, reference_dates, delays_unit)
}
\arguments{
\item{reporting_triangle_matrix}{Matrix of reporting triangle data.}

\item{reference_dates}{Date vector of reference dates.}

\item{delays_unit}{Character string specifying temporal granularity.}
}
\description{
Internal helper to validate the arguments passed to new_reporting_triangle.
}
\keyword{internal}
