% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_rep_tri_df}
\alias{.validate_rep_tri_df}
\title{Validate the reporting triangle data.frame}
\usage{
.validate_rep_tri_df(data, delays_unit)
}
\arguments{
\item{data}{Data.frame in long tidy form with reference dates, report dates,
and case counts, used to create a \code{reporting_triangle} object.}

\item{delays_unit}{Character string specifying the temporal granularity of
the delays. Options are \code{"days"}, \code{"weeks"}, \code{"months"}, \code{"years"}.
For the matrix method, this is simply passed as an item in the
\code{reporting_triangle} object which will later be used to create a nowcast
data.frame. For the data.frame method, this is used to compute the delay
in terms of the specified unit, and to expand the combinations of
reference dates and delays to the complete set of combinations in
the reporting triangle.  Default is \code{"days"}.}
}
\value{
NULL, invisibly
}
\description{
Checks for duplicate reference date report dates, missing
columns, report dates beyond the final reference date, and missing
combinations of delays and reports.
}
