% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_and_apply_delay.R
\name{estimate_and_apply_delay}
\alias{estimate_and_apply_delay}
\title{Estimate and apply delay from a reporting triangle}
\usage{
estimate_and_apply_delay(
  reporting_triangle,
  max_delay,
  n_history_delay = round(1.5 * max_delay)
)
}
\arguments{
\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{max_delay}{Integer indicating the maximum delay to estimate, in units
of the delay. The default is to use the whole reporting triangle,
\code{ncol(reporting_triangle) -1}.}

\item{n_history_delay}{Integer indicating number of reference times used for
delay estimation.}
}
\value{
\code{pt_nowcast_matrix} Matrix of point nowcasts
}
\description{
Estimate and apply delay from a reporting triangle
}
\examples{
triangle <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, NA,
    80, 40, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 5,
  byrow = TRUE
)
pt_nowcast_matrix<- estimate_and_apply_delay(
  reporting_triangle = triangle,
  max_delay = 3,
  n = 4
)
pt_nowcast_matrix
}
