% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_delay.R
\name{estimate_delay}
\alias{estimate_delay}
\title{Estimate a delay distribution from a reporting triangle}
\usage{
estimate_delay(
  triangle,
  max_delay = ncol(triangle) - 2,
  n_history = nrow(triangle)
)
}
\arguments{
\item{triangle}{dataframe of the reporting triangle, with one column for the
reference date and the remaining columns specifying the delay (units not
specified). Values indicate the number of new observations assigned to
the reference date at the specified delay.}

\item{max_delay}{integer indicating the maximum delay to estimate, in units
of the delay. The default is to use the whole reporting triangle,
\code{ncol(triangle) -1}.}

\item{n_history}{integer indicating the number of reference dates to be
used in the estimate of the reporting delay, always starting from the most
recent reporting delay. The default is to use the whole reporting triangle,
so \code{nrow(triangle)-1}}
}
\value{
delay_df dataframe of length \code{max_delay} with columns \code{delay}
and \code{pmf}, indicating the point estimate of the empirical probability
mass on each delay
}
\description{
Provides an estimate of the reporting delay as a function
of the delay, based on the reporting triangle and the specified maximum
delay and number of reference date observations to be used in the estimation.
This point estimate of the delay is computed empirically, using an
iterative algorithm starting from the most recent observations. It was
modified from the code originally developed by the Karlsruhe Institute
of Technology RESPINOW German Hospitalization Nowcasting Hub,
Modified from: https://github.com/KITmetricslab/RESPINOW-Hub/blob/7cce3ae2728116e8c8cc0e4ab29074462c24650e/code/baseline/functions.R#L55 #nolint
}
\examples{
library(epinowcast)
nat_germany_hosp <-
  germany_covid19_hosp[location == "DE"][age_group == "00+"]
nat_germany_hosp <- enw_filter_report_dates(
  nat_germany_hosp,
  latest_date = "2021-10-01"
)
pobs <- enw_preprocess_data(nat_germany_hosp, max_delay = 21)
triangle_raw <- pobs$reporting_triangle[[1]]
delay_df <- estimate_delay(triangle_raw[, -1],
  max_delay = 20,
  n_history = 30
)
}
