% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dispersion.R
\name{estimate_dispersion}
\alias{estimate_dispersion}
\title{Estimate dispersion parameters}
\usage{
estimate_dispersion(
  list_of_nowcasts,
  list_of_trunc_rts,
  n = length(list_of_nowcasts)
)
}
\arguments{
\item{list_of_nowcasts}{List of complete reporting rectangles where rows
represent reference time points and columns represent delays.}

\item{list_of_trunc_rts}{List of truncated reporting triangle matrices,
containing all observations as of the latest reference time. Elements of
list are paired with elements of \code{list_of_nowcasts}}

\item{n}{Integer indicating the number of reporting rectangles to use to
estimate the dispersion parameters}
}
\value{
Vector of length one less than the number of columns in the
latest reporting triangle, with each element representing the estimate
of the dispersion parameter for each delay d, starting at delay d=1.
}
\description{
This function ingests a list of point nowcasts (completed reporting
rectangles) and a corresponding list of truncated reporting triangles and
estimates and uses both to estimate a vector of negative binomial
dispersion parameters from the observations and estimates at each delay,
starting at delay = 1.
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

trunc_rts <- truncate_triangles(
  triangle = triangle,
  n = 2
)
retro_rts <- generate_triangles(
  list_of_trunc_rts = trunc_rts
)
retro_nowcasts <- generate_point_nowcasts(
  list_of_rts = retro_rts,
  n = 5
)
disp_params <- estimate_dispersion(
  list_of_nowcasts = retro_nowcasts,
  list_of_trunc_rts = trunc_rts,
  n = 2
)
}
