% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{estimate_uncertainty}
\alias{estimate_uncertainty}
\title{Estimate uncertainty parameters}
\usage{
estimate_uncertainty(
  point_nowcast_matrices,
  truncated_reporting_triangles,
  retro_reporting_triangles,
  n = length(point_nowcast_matrices),
  error_model = fit_distribution,
  error_args = list(observation_model_name = "negative binomial"),
  ref_time_aggregator = ref_time_aggregator_default(),
  delay_aggregator = delay_aggregator_default()
)
}
\arguments{
\item{point_nowcast_matrices}{List of point nowcast matrices where rows
represent reference time points and columns represent delays.}

\item{truncated_reporting_triangles}{List of truncated reporting matrices,
containing all observations as of the latest reference time. Elements of
list are paired with elements of \code{point_nowcast_matrices}.}

\item{retro_reporting_triangles}{List of \code{n} truncated reporting triangle
matrices with as many rows as available given the truncation.}

\item{n}{Integer indicating the number of reporting matrices to use to
estimate the dispersion parameters.}

\item{error_model}{Function that ingests a matrix of observations and a
matrix of predictions and returns a vector that can be used to
apply uncertainty using the same error model. Default is
fit_distribution.}

\item{error_args}{List of arguments needed for the specified error model.
Default is \verb{list(observation_model_name = "negative_binomial").}}

\item{ref_time_aggregator}{Function that operates along the rows (reference
times) of the retrospective point nowcast matrix before it has been
aggregated across columns (delays). Default is \code{zoo::rollsum}.}

\item{delay_aggregator}{Function that operates along the columns (delays)
of the retrospective point nowcast matrix after it has been aggregated
across reference times. Default is \code{rowSums}.}

\item{ref_time_aggregator_args}{List of arguments needed for the specified
\code{ref_time_aggregator}. Default is \code{list(k=1, align = "right")}.}

\item{delay_aggregator_args}{List of arguments needed for the specified
\code{delay_aggregator}. Default is \code{list(na.rm = TRUE)}.}
}
\value{
\code{uncertainty_params} Vector of length of the number of horizons,
with each element representing the estimate of the uncertainty parameter
for each horizon. The specific parameter type depends on the chosen error
model.
}
\description{
This function ingests a list of point nowcast matrices and a corresponding
list of truncated reporting matrices and uses both to estimate a
vector of uncertainty parameters from the observations
and estimates at each horizon, starting at 0 up until the max delay
number of horizons.
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

trunc_rts <- truncate_triangles(triangle, n = 2)
retro_rts <- construct_triangles(trunc_rts)

retro_nowcasts <- fill_triangles(retro_rts, n = 5)
# Estimate dispersion parameters using default (negative binomial error
# model on the sums
disp_params <- estimate_uncertainty(
  point_nowcast_matrices = retro_nowcasts,
  truncated_reporting_triangles = trunc_rts,
  retro_reporting_triangles = retro_rts
)
disp_params

# Estimate dispersion parameters from rolling mean with a normal error model
disp_params_agg <- estimate_uncertainty(
  point_nowcast_matrices = retro_nowcasts,
  truncated_reporting_triangles = trunc_rts,
  retro_reporting_triangles = retro_rts,
  n = 2,
  error_model = fit_distribution,
  error_args = list(observation_model_name = "normal"),
  ref_time_aggregator = ref_time_aggregator_default(
    fun = zoo::rollmean,
    k = 1,
    align = "right"
  )
)
disp_params_agg
}
