% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{estimate_uncertainty}
\alias{estimate_uncertainty}
\title{Estimate dispersion parameters}
\usage{
estimate_uncertainty(
  point_nowcast_matrices,
  truncated_reporting_triangles,
  retro_reporting_triangles,
  n = length(point_nowcast_matrices),
  fun_to_aggregate = sum,
  k = 1
)
}
\arguments{
\item{point_nowcast_matrices}{List of point nowcast matrices where rows
represent reference time points and columns represent delays.}

\item{truncated_reporting_triangles}{List of truncated reporting matrices,
containing all observations as of the latest reference time. Elements of
list are paired with elements of \code{point_nowcast_matrices}.}

\item{retro_reporting_triangles}{List of \code{n} truncated reporting triangle
matrices with as many rows as available given the truncation.}

\item{n}{Integer indicating the number of reporting matrices to use to
estimate the dispersion parameters.}

\item{fun_to_aggregate}{Function that will operate along the nowcast
vectors after summing across delays. Eventually, we can add things like
mean, but for now since we are only providing a negative binomial
observation model, we can only allow sum. Currently supported
functions: \code{sum}.}

\item{k}{Integer indicating the number of reference times to apply the
\code{fun_to_aggregate} over to create target used to compute the nowcast
errors.}
}
\value{
Vector of length one less than the number of columns in the
latest reporting triangle, with each element representing the estimate
of the dispersion parameter for each delay d, starting at delay d=1.
}
\description{
This function ingests a list of point nowcast matrices and a corresponding
list of truncated reporting matrices and uses both to estimate a
vector of negative binomial dispersion parameters from the observations
and estimates at each horizon, starting at 0 up until the max delay
number of horizons.
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

trunc_rts <- truncate_triangles(triangle, n = 2)
retro_rts <- construct_triangles(trunc_rts)

retro_nowcasts <- fill_triangles(retro_rts, n = 5)
disp_params <- estimate_uncertainty(
  point_nowcast_matrices = retro_nowcasts,
  truncated_reporting_triangles = trunc_rts,
  retro_reporting_triangles = retro_rts,
  n = 2
)
disp_params

# Estimate dispersion parameters from rolling sum
disp_params_agg <- estimate_uncertainty(
  point_nowcast_matrices = retro_nowcasts,
  truncated_reporting_triangles = trunc_rts,
  retro_reporting_triangles = retro_rts,
  n = 2,
  fun_to_aggregate = sum,
  k = 2
)
disp_params_agg
}
