% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_retro_triangles.R
\name{generate_retro_triangles}
\alias{generate_retro_triangles}
\title{Generate retrospective reporting triangles}
\usage{
generate_retro_triangles(
  triangle,
  n_triangles = nrow(triangle) - ncol(triangle)
)
}
\arguments{
\item{triangle}{Matrix of the reporting triangle
to be used to generate retrospective nowcasts, with rows representing the
time points of reference and columns representing the delays}

\item{n_triangles}{Integer indicating the number of retrospective
reporting triangles to be generated, always starting from the most
recent reference time. Default is  only generate truncated matrices
that have sufficient number of rows to generate a nowcast from, though
any number can be specified.}
}
\value{
List of \code{n_triangle} retrospective reporting triangle matrices
with as many rows as available given the truncation, and the same number
of columns as \code{triangle}
}
\description{
This function ingests a reporting triangle and the number of retrospective
reporting triangles we want to create, and iteratively generates the
reporting triangle that would have been available as of the maximum
reference time, and iteratively working backwards for
\code{n_history_uncertainty} snapshots
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

retro_rts <- generate_retro_triangles(
  triangle = triangle,
  n_triangles = 2
)
print(retro_rts[[1]])
print(retro_rts[[2]])
}
