% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_neg_vals.R
\name{handle_neg_vals}
\alias{handle_neg_vals}
\title{Handle negative values in the reporting triangle}
\usage{
handle_neg_vals(triangle)
}
\arguments{
\item{triangle}{the reporting triangle as a matrix, where rows are the
time points and columns are the delays, already truncated to the maximum
delay and the number of historical observations}
}
\value{
pos_triangle a positive integer matrix with negative values of
reporting handled via passing them to the subsequent days delay
}
\description{
Takes in a reporting triangle and returns a matrix in the same format
as the input triangle, but with negative values of reporting handled via passing
them to the subsequent days delay.
Modified from https://github.com/KITmetricslab/RESPINOW-Hub/blob/main/code/baseline/functions.R #nolint
}
