% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-methods.R
\name{head.reporting_triangle}
\alias{head.reporting_triangle}
\title{Get first rows of a reporting_triangle}
\usage{
\method{head}{reporting_triangle}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \link{reporting_triangle} object.}

\item{n}{Integer indicating the number of rows to return. Default is 6.}

\item{...}{Additional arguments (not currently used).}
}
\value{
First rows as a reporting_triangle.
}
\description{
Get first rows of a reporting_triangle
}
\seealso{
Reporting triangle construction and validation
\code{\link{[<-.reporting_triangle}()},
\code{\link{as.data.frame.reporting_triangle}()},
\code{\link{as.matrix.reporting_triangle}()},
\code{\link{as_ChainLadder_triangle}()},
\code{\link{as_reporting_triangle}()},
\code{\link{as_reporting_triangle.data.frame}()},
\code{\link{as_reporting_triangle.matrix}()},
\code{\link{as_reporting_triangle.triangle}()},
\code{\link{assert_reporting_triangle}()},
\code{\link{get_delay_from_dates_function}()},
\code{\link{get_delay_unit_function}()},
\code{\link{get_delays_unit}()},
\code{\link{get_max_delay}()},
\code{\link{get_mean_delay}()},
\code{\link{get_quantile_delay}()},
\code{\link{get_reference_dates}()},
\code{\link{get_reporting_structure}()},
\code{\link{is_reporting_triangle}()},
\code{\link{new_reporting_triangle}()},
\code{\link{print.reporting_triangle}()},
\code{\link{reporting_triangle-class}},
\code{\link{summary.reporting_triangle}()},
\code{\link{tail.reporting_triangle}()},
\code{\link{truncate_to_delay}()},
\code{\link{truncate_to_quantile}()},
\code{\link{validate_reporting_triangle}()}
}
\concept{reporting_triangle}
