% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinenowcast.R
\name{new_nowcast_df}
\alias{new_nowcast_df}
\title{Combine data from a nowcast dataframe, strata, and reference dates}
\usage{
new_nowcast_df(nowcast_df, strata_map, reference_dates)
}
\arguments{
\item{nowcast_df}{Data.frame containing information for multiple draws with
columns for the reference time (\code{time}), the predicted counts
(\code{pred_count}), and optionally the draw number (\code{draw}).}

\item{strata_map}{Named list where each name is the grouping
variable (e.g. \code{age_group}) and each value is the stratum (e.g. "0-18").}

\item{reference_dates}{Vector of reference dates corresponding to the
reference times in the \code{nowcast_df}.}
}
\value{
An object of class \code{\link{nowcast_df}}
}
\description{
Combines data from a nowcast dataframe, a named list of the
strata associated with the nowcast dataframe, and a vector of reference
dates corresponding to the time column in the \code{nowcast_df}
}
