% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{new_reporting_triangle}
\alias{new_reporting_triangle}
\title{Class constructor for \code{reporting_triangle} objects}
\usage{
new_reporting_triangle(data)
}
\arguments{
\item{data}{A list to convert}
}
\value{
An object of class \code{reporting_triangle}
}
\description{
Class constructor for \code{reporting_triangle} objects
}
