% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_reporting_triangle.R
\name{new_reporting_triangle}
\alias{new_reporting_triangle}
\title{Class constructor for \code{reporting_triangle} objects}
\usage{
new_reporting_triangle(
  reporting_triangle_matrix,
  reference_dates,
  structure,
  max_delay,
  delays_unit,
  strata = NULL
)
}
\arguments{
\item{reporting_triangle_matrix}{Matrix of reporting triangle}

\item{reference_dates}{Vector of character strings indicating the reference
dates corresponding to each row of the reporting triangle matrix (\code{data}).}

\item{structure}{Integer or vector specifying the reporting structure.
If integer, divides columns evenly by that integer (with last possibly
truncated).  If vector, the sum must not be greater than or equal to the
number of columns. Default is 1 (standard triangular structure).}

\item{max_delay}{Integer indicating the maximum delay.}

\item{delays_unit}{Character string specifying the temporal granularity of
the delays. Options are \code{"days"}, \code{"weeks"}, \code{"months"}, \code{"years"}.
For the matrix method, this is simply passed as an item in the
\code{reporting_triangle} object which will later be used to create a nowcast
data.frame. For the data.frame method, this is used to compute the delay
in terms of the specified unit, and to expand the combinations of
reference dates and delays to the complete set of combinations in
the reporting triangle.  Default is \code{"days"}.}

\item{strata}{Named list indicating where the name will correspond to
the column name and the single entry will correspond to the variable within
the column.}
}
\value{
An object of class \code{reporting_triangle}
}
\description{
Class constructor for \code{reporting_triangle} objects
}
