% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_lower_right_with_NA}
\alias{replace_lower_right_with_NA}
\title{Replace the lower right triangle of the matrix with NAs}
\usage{
replace_lower_right_with_NA(matrix)
}
\arguments{
\item{matrix}{A matrix}
}
\value{
A mattrix of the same dimensions, with NAs for all the lower right
entries
}
\description{
Replace the lower right triangle of the matrix with NAs
}
