% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-class.R
\name{reporting_triangle-class}
\alias{reporting_triangle-class}
\alias{reporting_triangle}
\title{Reporting Triangle Object}
\description{
A \code{reporting_triangle} object contains the data and metadata needed for
nowcasting.
}
\section{Structure}{

A \code{reporting_triangle} is a list with the following components:
\describe{
\item{reporting_triangle_matrix}{Matrix with which rows are reference
times and columns are delays and
entries are incident cases at each reference time and delay.}
\item{reference_dates}{Vector of the same length as the rows of the
matrix indicating the dates corresponding to the reference times in the
rows of the reporting triangle.}
\item{structure}{Vector indicating the "structure" of the reporting
triangle, see \code{\link[=construct_triangle]{construct_triangle()}} for more details.}
\item{max_delay}{Integer indicating the maximum delay.}
\item{delays_unit}{Character string indicating the unit of the delays.
Valid options are "days", "weeks", "months", "years".}
\item{strata}{Character string indicating the strata.}
}
See the corresponding \code{\link[=as_reporting_triangle.matrix]{as_reporting_triangle.matrix()}} and
\code{\link[=as_reporting_triangle.data.frame]{as_reporting_triangle.data.frame()}} functions
for more details on the required input formats to generate the object.
}

\seealso{
Reporting triangle construction and validation
\code{\link{as_ChainLadder_triangle}()},
\code{\link{as_reporting_triangle}()},
\code{\link{as_reporting_triangle.data.frame}()},
\code{\link{as_reporting_triangle.matrix}()},
\code{\link{as_reporting_triangle.triangle}()},
\code{\link{assert_reporting_triangle}()},
\code{\link{detect_structure}()},
\code{\link{new_reporting_triangle}()}
}
\concept{reporting_triangle}
