% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distribution.R
\name{sample_distribution}
\alias{sample_distribution}
\title{Sample from observations model given predictions}
\usage{
sample_distribution(
  pred,
  uncertainty_params,
  observation_model_name = "negative binomial"
)
}
\arguments{
\item{pred}{Matrix or vector of predictions.}

\item{uncertainty_params}{Vector of uncertainty parameters.}

\item{observation_model_name}{Character string indicating the choice of
observation model to sample to the predicted nowcasts.
Default is \code{"negative binomial"}.}
}
\value{
\code{sampled_pred} Object of the same dimensions as \code{pred} representing
a single draw from the \code{observation_model_name} distribution
with the specified \verb{uncertainty params}.
}
\description{
Sample from observations model given predictions
}
\examples{
pred <- matrix(c(
  3.7, 6.1,
  5.2, 10.4
), byrow = 2, nrow = 2)
sampled_preds <- sample_distribution(pred,
  uncertainty_params = c(50, 100),
  observation_model_name = "negative binomial"
)
sampled_preds
}
