% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_predictions.R
\name{sample_nowcast}
\alias{sample_nowcast}
\title{Generate a single draw of a nowcast combining observed and predicted values}
\usage{
sample_nowcast(point_nowcast_matrix, reporting_triangle, dispersion, ...)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of point nowcast predictions and
observations, with rows representing the reference times and columns
representing the delays.}

\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{dispersion}{Vector of dispersion parameters indexed by horizon from
minus one to the maximum delay.}

\item{...}{Additional arguments passed to \code{sample_predictions()}}
}
\value{
Vector of predicted counts at each reference time based on combining
the observed counts and the predicted counts for the unobserved elements.
}
\description{
Generate a single draw of a nowcast combining observed and predicted values
}
\examples{
point_nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 16.8,
    80, 40, 21.2, 19.5,
    70, 34.5, 15.4, 9.1
  ),
  nrow = 5,
  byrow = TRUE
)
reporting_triangle <- construct_triangle(point_nowcast_matrix)
disp <- c(0.8, 12.4, 9.1)
nowcast_draw <- sample_nowcast(
  point_nowcast_matrix,
  reporting_triangle,
  disp
)
nowcast_draw
}
