% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_predictions.R
\name{sample_nowcasts}
\alias{sample_nowcasts}
\title{Generate multiple draws of a nowcast combining observed and predicted values}
\usage{
sample_nowcasts(
  point_nowcast_matrix,
  reporting_triangle,
  uncertainty_params,
  draws = 1000,
  error_model = sample_distribution,
  error_args = list(observation_model_name = "negative binomial"),
  aggregator = zoo::rollsum,
  aggregator_args = list(k = 1, align = "right")
)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of point nowcast predictions and
observations, with rows representing the reference times and columns
representing the delays.}

\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{uncertainty_params}{Vector of uncertainty parameters ordered from
horizon 1 to the maximum horizon. Note that these will be reversed
internally to match the ordering of the \code{point_nowcast_matrix} (where
a horizon of 1 is the last entry).}

\item{draws}{Integer indicating the number of draws of the predicted
nowcast vector to generate. Default is \code{1000}.}

\item{error_model}{Function that ingests a matrix of predictions and a vector
of uncertainty parameters and generates draws from the error model.
Default is \code{sample_distribution}.}

\item{error_args}{List of arguments needed for the specified error model.
Default is \verb{list(observation_model_name = "negative_binomial").}}
}
\value{
Dataframe containing information for multiple draws with columns
for the reference time (\code{time}), the predicted counts (\code{pred_count}), and
the draw number (\code{draw}).
}
\description{
Generate multiple draws of a nowcast combining observed and predicted values
}
\examples{
point_nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 16.8,
    80, 40, 21.2, 19.5,
    70, 34.5, 15.4, 9.1
  ),
  nrow = 5,
  byrow = TRUE
)
reporting_triangle <- construct_triangle(point_nowcast_matrix)
disp <- c(0.8, 12.4, 9.1)
nowcast_draws <- sample_nowcasts(
  point_nowcast_matrix,
  reporting_triangle,
  disp,
  draws = 5
)
nowcast_draws
}
