% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_prediction.R, R/sample_predictions.R
\name{sample_prediction}
\alias{sample_prediction}
\title{Get a draw of only the predicted elements of the nowcast vector}
\usage{
sample_prediction(
  point_nowcast_matrix,
  reporting_triangle,
  uncertainty_params,
  error_model = sample_distribution,
  error_args = list(observation_model_name = "negative binomial"),
  aggregator = zoo::rollsum,
  aggregator_args = list(k = 1, align = "right")
)

sample_prediction(
  point_nowcast_matrix,
  reporting_triangle,
  uncertainty_params,
  error_model = sample_distribution,
  error_args = list(observation_model_name = "negative binomial"),
  aggregator = zoo::rollsum,
  aggregator_args = list(k = 1, align = "right")
)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of point nowcast predictions and
observations, with rows representing the reference times and columns
representing the delays.}

\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{uncertainty_params}{Vector of uncertainty parameters ordered from
horizon 1 to the maximum horizon.}

\item{error_model}{Function that ingests a matrix of predictions and a vector
of uncertainty parameters and generates draws from the error model.
Default is \code{sample_distribution}.}

\item{error_args}{List of arguments needed for the specified error model.
Default is \verb{list(observation_model_name = "negative_binomial").}}

\item{aggregator}{Function that operates along the rows of the retrospective
point nowcast matrix after it has been aggregated across columns (delays)
Default is \code{zoo::rollsum}.}

\item{aggregator_args}{List of arguments needed for the specified aggregator.
Default is \code{list(k=1, align = "right")}.}
}
\value{
Vector of predicted draws at each reference time, for all reference
times in the input \code{point_nowcast_matrix}.

Vector of predicted draws at each reference time, for all reference
times in the input \code{point_nowcast_matrix}.
}
\description{
Get a draw of only the predicted elements of the nowcast vector

Get a draw of only the predicted elements of the nowcast vector
}
\examples{
point_nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 16.8,
    80, 40, 21.2, 19.5,
    70, 34.5, 15.4, 9.1
  ),
  nrow = 5,
  byrow = TRUE
)
reporting_triangle <- construct_triangle(point_nowcast_matrix)
disp <- c(0.8, 12.4, 9.1)
nowcast_pred_draw <- sample_prediction(
  point_nowcast_matrix,
  reporting_triangle,
  disp
)
nowcast_pred_draw

# Get draws on the rolling sum
nowcast_pred_draw_agg <- sample_prediction(
  point_nowcast_matrix,
  reporting_triangle,
  disp,
  fun_to_aggregate = sum,
  k = 2
)
nowcast_pred_draw_agg
point_nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 16.8,
    80, 40, 21.2, 19.5,
    70, 34.5, 15.4, 9.1
  ),
  nrow = 5,
  byrow = TRUE
)
reporting_triangle <- construct_triangle(point_nowcast_matrix)
disp <- c(0.8, 12.4, 9.1)
nowcast_pred_draw <- sample_prediction(
  point_nowcast_matrix,
  reporting_triangle,
  disp
)
nowcast_pred_draw

# Get draws on the rolling sum
nowcast_pred_draw_agg <- sample_prediction(
  point_nowcast_matrix,
  reporting_triangle,
  disp,
  aggregator = zoo::rollsum,
  aggregator_args = list(k = 2, align = "right")
)
nowcast_pred_draw_agg
}
