% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_predictions.R
\name{sample_predictions}
\alias{sample_predictions}
\title{Get a dataframe of multiple draws of only the predicted elements of the
nowcast vector}
\usage{
sample_predictions(
  point_nowcast_matrix,
  reporting_triangle,
  uncertainty_params,
  draws = 1000,
  error_model = function(pred, params) sample_nb(pred, params),
  ref_time_aggregator = function(x) identity(x),
  delay_aggregator = function(x) rowSums(x, na.rm = TRUE)
)
}
\arguments{
\item{point_nowcast_matrix}{Matrix of point nowcast predictions and
observations, with rows representing the reference times and columns
representing the delays.}

\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{uncertainty_params}{Vector of uncertainty parameters ordered from
horizon 1 to the maximum horizon. Note that these will be reversed
internally to match the ordering of the \code{point_nowcast_matrix} (where
a horizon of 1 is the last entry).}

\item{draws}{Integer indicating the number of draws of the predicted
nowcast vector to generate. Default is \code{1000}.}

\item{error_model}{Function that ingests a vector or matrix of predictions
and a vector of uncertainty parameters and generates draws from the
error model. Default is \code{sample_nb()}.}

\item{ref_time_aggregator}{Function that operates along the rows (reference
times) of the retrospective point nowcast matrix before it has been
aggregated across columns (delays). Default is \code{function(x) identity(x)}
which does not aggregate across refrence times.}

\item{delay_aggregator}{Function that operates along the columns (delays)
of the retrospective point nowcast matrix after it has been aggregated
across reference times. Default is \code{function(x) rowSums(x, na.rm = TRUE)}.}
}
\value{
Dataframe containing the predicted point nowcast vectors indexed by
reference time (\code{pred_count}), reference time (\code{time}), and the draw index
(\code{draw}).
}
\description{
Get a dataframe of multiple draws of only the predicted elements of the
nowcast vector
}
\examples{
point_nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 16.8,
    80, 40, 21.2, 19.5,
    70, 34.5, 15.4, 9.1
  ),
  nrow = 5,
  byrow = TRUE
)
reporting_triangle <- construct_triangle(point_nowcast_matrix)
disp <- c(0.8, 12.4, 9.1)
nowcast_pred_draws <- sample_predictions(
  point_nowcast_matrix,
  reporting_triangle,
  disp,
  draws = 5
)
nowcast_pred_draws
# Get nowcast pred draws over rolling sum
nowcast_pred_draws_rolling_df <- sample_predictions(
  point_nowcast_matrix,
  reporting_triangle,
  disp,
  500,
  ref_time_aggregator = function(x) zoo::rollsum(x, k = 2, align = "right")
)
nowcast_pred_draws_rolling_df
}
