% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_triangle-methods.R
\name{[<-.reporting_triangle}
\alias{[<-.reporting_triangle}
\title{Subset assignment for reporting_triangle objects}
\usage{
\method{[}{reporting_triangle}(x, ...) <- value
}
\arguments{
\item{x}{A \link{reporting_triangle} object.}

\item{...}{Row and column indices for assignment.}

\item{value}{Values to assign.}
}
\value{
The modified reporting_triangle object.
}
\description{
Assignment method that allows modification of reporting_triangle values
while validating the result remains a valid reporting_triangle.
}
\seealso{
Reporting triangle construction and validation
\code{\link{[.reporting_triangle}()},
\code{\link{as.data.frame.reporting_triangle}()},
\code{\link{as.matrix.reporting_triangle}()},
\code{\link{as_ChainLadder_triangle}()},
\code{\link{as_reporting_triangle}()},
\code{\link{as_reporting_triangle.data.frame}()},
\code{\link{as_reporting_triangle.matrix}()},
\code{\link{as_reporting_triangle.triangle}()},
\code{\link{assert_reporting_triangle}()},
\code{\link{get_delays_from_dates}()},
\code{\link{get_delays_unit}()},
\code{\link{get_max_delay}()},
\code{\link{get_mean_delay}()},
\code{\link{get_quantile_delay}()},
\code{\link{get_reference_dates}()},
\code{\link{get_report_dates}()},
\code{\link{get_reporting_structure}()},
\code{\link{head.reporting_triangle}()},
\code{\link{is_reporting_triangle}()},
\code{\link{new_reporting_triangle}()},
\code{\link{print.reporting_triangle}()},
\code{\link{reporting_triangle-class}},
\code{\link{summary.reporting_triangle}()},
\code{\link{tail.reporting_triangle}()},
\code{\link{truncate_to_delay}()},
\code{\link{truncate_to_quantile}()},
\code{\link{validate_reporting_triangle}()}
}
\concept{reporting_triangle}
