% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_triangles.R
\name{truncate_triangle}
\alias{truncate_triangle}
\title{Get a single truncated triangle}
\usage{
truncate_triangle(t, reporting_triangle)
}
\arguments{
\item{t}{Integer indicating the number of timepoints to truncate off the
bottom of the original reporting triangle.}

\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}
}
\value{
\code{trunc_rep_tri} Matrix with \code{t} fewer rows than
\code{reporting_triangle}.
}
\description{
This function takes in a integer \code{t} and a reporting triangle and generates
a truncated reporting triangle, remove the last \code{t} observations.
}
\examples{
# example code
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

trunc_rep_tri <- truncate_triangle(t = 1, reporting_triangle = triangle)
trunc_rep_tri
}
\seealso{
Other generate_retrospective_data: 
\code{\link{generate_triangle}()},
\code{\link{generate_triangles}()},
\code{\link{truncate_triangles}()}
}
\concept{generate_retrospective_data}
