% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_triangles.R
\name{truncate_triangles}
\alias{truncate_triangles}
\title{Generate truncated reporting triangles}
\usage{
truncate_triangles(
  reporting_triangle,
  n = min(ncol(reporting_triangle) - 1, nrow(reporting_triangle) - 1)
)
}
\arguments{
\item{reporting_triangle}{Matrix of the reporting triangle, with rows
representing the time points of reference and columns representing the
delays. Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row. (e.g. weekly reporting of daily data).}

\item{n}{Integer indicating the number of retrospective
truncated triangles to be generated, always starting from the most
recent reference time. Default is to generate truncated matrices for
the maximum delay (number of columns - 1) or one less than the number of
available rows.}
}
\value{
\code{trunc_rep_tri_list} List of \code{n} truncated reporting triangle
matrices with as many rows as available given the truncation, and the same
number of columns as \code{reporting_triangle}.
}
\description{
This function ingests a reporting triangle/matrix and the number of
truncated reporting triangles we want to create, \code{n}, and iteratively
truncates the reporting triangle, working from the latest reference time
(bottom) to the older reference times (top) for \code{n}
snapshots.
}
\examples{
triangle <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)

truncated_rts <- truncate_triangles(triangle, n = 2)
truncated_rts[1:2]
}
