% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_triangles.R
\name{truncate_triangles}
\alias{truncate_triangles}
\title{Generate truncated reporting triangles}
\usage{
truncate_triangles(
  reporting_triangle,
  n = nrow(reporting_triangle) - sum(is.na(rowSums(reporting_triangle))) - 1
)
}
\arguments{
\item{reporting_triangle}{A \link{reporting_triangle} object with rows
representing reference times and columns representing delays.
Can be a reporting matrix or incomplete reporting matrix.
Can also be a ragged reporting triangle, where multiple columns are
reported for the same row (e.g., weekly reporting of daily data).}

\item{n}{Integer indicating the number of retrospective
truncated triangles to be generated, always starting from the most
recent reference time. Default is to generate truncated matrices for
each row up until there are insufficient rows to generate nowcasts
from, where the minimum requirement is one more than the  number of
horizon rows (rows containing NAs).}
}
\value{
\code{trunc_rep_tri_list} List of \code{n} truncated \code{reporting_triangle}
objects with as many rows as available given the truncation, and the same
number of columns as the input \code{reporting_triangle}.
}
\description{
This function ingests a reporting triangle/matrix and the number of
truncated reporting triangles we want to create, \code{n}, and iteratively
truncates the reporting triangle, working from the latest reference time
(bottom) to the older reference times (top) for \code{n}
snapshots.
}
\examples{
triangle_mat <- matrix(
  c(
    65, 46, 21, 7,
    70, 40, 20, 5,
    80, 50, 10, 10,
    100, 40, 31, 20,
    95, 45, 21, NA,
    82, 42, NA, NA,
    70, NA, NA, NA
  ),
  nrow = 7,
  byrow = TRUE
)
ref_dates <- seq(as.Date("2025-01-01"), by = "day", length.out = 7)
triangle <- as_reporting_triangle(
  data = triangle_mat,
  reference_dates = ref_dates,
  max_delay = 3
)

truncated_rts <- truncate_triangles(triangle, n = 2)
truncated_rts[1:2]
}
\seealso{
Retrospective data generation functions
\code{\link{construct_triangle}()},
\code{\link{construct_triangles}()},
\code{\link{truncate_triangle}()}
}
\concept{generate_retrospective_data}
