% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_triangle}
\alias{validate_triangle}
\title{Validate triangle}
\usage{
validate_triangle(triangle)
}
\arguments{
\item{triangle}{a dataframe formatted as a reporting triangle}
}
\value{
NULL, invisibly
}
\description{
Various checks to make sure that the reporting triangle passed in to
\code{estimate_delay()} is formatted properly.
}
